rust: remove obsolete table wire surface, retain cook and block accelerators
Prune the pre-id-table Rust table wire and reflection descriptors:
- Remove codecs.go, descriptors.go, and runtime.go from internal/codegen/rusttable.
- Retain cook and block accelerators, compile-time layout asserts, and build version check.
- Prune tablenames/rust.go to the 14 retained accelerator/version identifiers.
- Update conformance driver to run cook and block surfaces only; remove dead wire and soak code.
- Remove obsolete wire targets (tables-rust-walk, tables-rust-soak, tables-rust-alloc-audit) from make/rust.mk.
- Update ROADMAP.md and docs/PORTING.md to reflect table wire removal deferred to #518.
- Update test/conformance/rust/Cargo.toml to depend only on graphdemo and blockdemo.

Author: Emma Gemini
