+++ compiler/issue715_test.go
@@ -281,7 +281,7 @@ class Program {
 
 	t.Run("emitted_text", func(t *testing.T) {
 		uNoDefault := unitFromSource(t, "package p\ntable Ship { tag bytes(4)\n after int32 }\n")
-		for _, lang := range []string{"cpp", "c", "go", "cs", "java", "js", "dart", "rust", "elixir"} {
+		for _, lang := range []string{"cpp", "c", "go", "cs", "java", "js", "dart", "elixir"} {
 			files, err := c.Generate(uNoDefault, lang, Options{})
 			if err != nil {
 				t.Fatalf("generate %s: %v", lang, err)
