--- internal/codegen/rusttable/runtime.go
@@ -1,2725 +1,0 @@
-package rusttable
-
-import (
-	"fmt"
-	"sort"
-	"strings"
-
-	"github.com/mas-bandwidth/schema/v2/ir"
-)
-
-// runtimeModule assembles table_runtime.rs — the unit's ONE shared table
-// runtime: the report, the writer and the reader, the enum-keyed array's
-// storage, the reflection descriptors, and the text form's single generic
-// walk over them (docs/SPEC-TABLES.md §16.1).
-//
-// It is emitted ONCE PER UNIT and its body is the same bytes in every unit,
-// whatever that unit declares — the Rust twin of the generic-walk gate the
-// C++ backend is held to. Nothing in it names a table, a package or a kind
-// this unit happens to use.
-func runtimeModule(u *ir.Unit, closure map[string]bool) []byte {
-	var b strings.Builder
-	b.WriteString(header(runtimeSourceName(u), u.Package, "the TABLE wire's shared runtime (docs/SPEC-TABLES.md)"))
-	b.WriteString(runtimeBanner)
-	b.WriteString(runtimeSource)
-	return []byte(b.String())
-}
-
-// runtimeSourceName is the file the runtime is attributed to: the unit's
-// first file that declares a table, so the banner names a real schema file
-// rather than the unit key.
-func runtimeSourceName(u *ir.Unit) string {
-	names := make([]string, 0, len(u.Files))
-	for _, f := range u.Files {
-		if len(f.Tables) > 0 {
-			names = append(names, f.Base)
-		}
-	}
-	sort.Strings(names)
-	if len(names) == 0 {
-		return u.Package
-	}
-	return names[0]
-}
-
-const runtimeBanner = `//
-// THE SHARED TABLE RUNTIME, emitted once per unit: the read report, the wire
-// writer and reader, the enum-keyed array's storage, the reflection
-// descriptors (§8), and the text form's ONE generic walk over them (§16.1).
-// Its bytes do not vary with what a unit declares — a port's twin of the
-// generic-walk gate the C++ backend is held to.
-//
-// Nothing here allocates: the caller owns the value, the buffer and the
-// report. There is no serialize dependency, so a table module compiles into
-// any crate.
-
-#![allow(clippy::needless_range_loop)]
-#![allow(clippy::too_many_arguments)]
-#![allow(clippy::manual_range_contains)]
-#![allow(clippy::collapsible_else_if)]
-#![allow(clippy::comparison_chain)]
-
-use core::ptr;
-
-`
-
-// runtimeSource is the whole Rust runtime. It is source, not a template: no
-// substitution happens, which is what makes "the same bytes in every unit"
-// true by construction.
-const runtimeSource = `// ---- the read report (docs/SPEC-TABLES.md §4) ----
-
-// The table-wire read report — the permissive contract's ledger. Silence
-// (the default) means the data matched this reader's schema exactly.
-#[derive(Clone, Copy, PartialEq, Eq, Debug, Default)]
-pub struct TableReport {
-    pub unknown: i32,      // unknown field ids skipped (newer data)
-    pub kind_mismatch: i32, // known id, changed type — skipped, never misdecoded
-    pub clamped: i32,      // out-of-range values clamped to declared bounds
-    // duplicate is the TEXT FORM's counter and the WIRE NEVER RAISES IT
-    // (docs/SPEC-TABLES.md §4, §16.2): a body carrying an id twice is legal input
-    // whose last occurrence wins, silently.
-    pub duplicate: i32,
-    pub malformed: bool, // framing damage; decode stopped, partial result kept
-}
-
-// ---- an enum's TABLE-wire identity (docs/SPEC-TABLES.md §5) ----
-
-// A value rides as the u16 hash of its VARIANT NAME, so a variant may be
-// added anywhere, removed, or reordered and old data still reads. None is the
-// one reserved id, 0, which no declared name can fold to.
-//
-// C++ and C# spell this as an overload set; Rust has no overloading, so it is
-// a trait the generated code implements once per enum. One unit-level name
-// (§11) rather than one per declaration.
-pub trait TableEnum: Copy + Sized {
-    // this value's wire id, or None when no variant names it — a value with
-    // no wire identity, which measure and save both refuse.
-    fn table_id(self) -> Option<u16>;
-    // the value a wire id names, or None when this build cannot name it.
-    fn table_value(id: u16) -> Option<Self>;
-}
-
-// ---- the enum-keyed array's storage (docs/SPEC-TABLES.md §2.4) ----
-
-// E.Max slots, ONE PER NAMED VARIANT, with the key k at index k - 1: the
-// storage SHIFTS LEFT and nothing is stored for None.
-//
-// NOTHING OUTSIDE THE ARRAY NAMES ITS SIZE. The extent is the const generic
-// the declaration writes as the key enum's own MAX, so there is no
-// constructor argument and no constant a consumer could put out of step
-// with it.
-//
-// NONE IS THE NULL KEY: it names no slot, it never rides on the wire, a
-// stored key of 0 is malformed, and INDEXING BY IT IS AN ERROR — a panic,
-// which stands in every build exactly as the C++ abort does.
-//
-// slots is public and is what the generated codecs walk, by STORAGE INDEX;
-// Index/IndexMut take the KEY, and they are the one place the None key is
-// caught.
-#[repr(C)]
-#[derive(Clone, Copy, PartialEq, Debug)]
-pub struct TableKeyed<T, const N: usize> {
-    pub slots: [T; N],
-}
-
-impl<T, const N: usize> TableKeyed<T, N> {
-    // the KEY a storage slot holds: slot i holds key i + 1 (§2.4). This is
-    // the one place the shift is spelled.
-    #[inline]
-    pub fn key_of_slot(slot: usize) -> u64 {
-        slot as u64 + 1
-    }
-
-    #[inline]
-    pub fn slot_of_key(key: u64) -> usize {
-        if key == 0 {
-            panic!("None is the null key of an enum-keyed array: it keys no slot");
-        }
-        key as usize - 1
-    }
-
-    // one slot: the KEY it holds and a reference to its element, so a walk
-    // of the whole array never spells a bound or the shift.
-    pub fn iter(&self) -> impl Iterator<Item = (u64, &T)> {
-        self.slots
-            .iter()
-            .enumerate()
-            .map(|(i, e)| (Self::key_of_slot(i), e))
-    }
-
-    pub fn iter_mut(&mut self) -> impl Iterator<Item = (u64, &mut T)> {
-        self.slots
-            .iter_mut()
-            .enumerate()
-            .map(|(i, e)| (Self::key_of_slot(i), e))
-    }
-}
-
-impl<T, const N: usize> core::ops::Index<u64> for TableKeyed<T, N> {
-    type Output = T;
-    fn index(&self, key: u64) -> &T {
-        &self.slots[Self::slot_of_key(key)]
-    }
-}
-
-impl<T, const N: usize> core::ops::IndexMut<u64> for TableKeyed<T, N> {
-    fn index_mut(&mut self, key: u64) -> &mut T {
-        &mut self.slots[Self::slot_of_key(key)]
-    }
-}
-
-impl<T: Copy + Default, const N: usize> Default for TableKeyed<T, N> {
-    fn default() -> Self {
-        TableKeyed {
-            slots: [T::default(); N],
-        }
-    }
-}
-
-// ---- the IEEE-754 bit patterns the wire carries for f32 and f64 (§3) ----
-
-#[inline]
-pub fn table_bits_to_float(bits: u32) -> f32 {
-    f32::from_bits(bits)
-}
-
-#[inline]
-pub fn table_float_to_bits(value: f32) -> u32 {
-    value.to_bits()
-}
-
-#[inline]
-pub fn table_bits_to_double(bits: u64) -> f64 {
-    f64::from_bits(bits)
-}
-
-#[inline]
-pub fn table_double_to_bits(value: f64) -> u64 {
-    value.to_bits()
-}
-
-// ---- the wire writer and reader ----
-
-// TableWriter writes the wire IN PLACE, into the caller's slice. Nothing
-// here allocates.
-//
-// EVERY PRIMITIVE IS FORCE-INLINED, and so is every fixed-class body that
-// calls one (#343): out of line, a writer's cursor round-trips through memory
-// at every put, and the two halves do not add up but multiply — only a body
-// with no call boundary in it lets the cursor stay in registers and lets
-// adjacent constant framing bytes merge into one store. The C++ reference
-// carries the same rule, spelled as a per-package macro because it has three
-// compilers to name; Rust has one attribute.
-pub struct TableWriter<'a> {
-    pub buffer: &'a mut [u8],
-    pub offset: usize,
-    pub overflow: bool,
-}
-
-impl<'a> TableWriter<'a> {
-    #[inline(always)]
-    pub fn new(buffer: &'a mut [u8]) -> TableWriter<'a> {
-        TableWriter {
-            buffer,
-            offset: 0,
-            overflow: false,
-        }
-    }
-
-    #[inline(always)]
-    pub fn raw(&mut self, data: &[u8]) {
-        if self.offset + data.len() > self.buffer.len() {
-            self.overflow = true;
-            return;
-        }
-        self.buffer[self.offset..self.offset + data.len()].copy_from_slice(data);
-        self.offset += data.len();
-    }
-
-    #[inline(always)]
-    pub fn put8(&mut self, v: u8) {
-        if self.offset + 1 > self.buffer.len() {
-            self.overflow = true;
-            return;
-        }
-        self.buffer[self.offset] = v;
-        self.offset += 1;
-    }
-
-    #[inline(always)]
-    pub fn put16(&mut self, v: u16) {
-        if self.offset + 2 > self.buffer.len() {
-            self.overflow = true;
-            return;
-        }
-        self.buffer[self.offset..self.offset + 2].copy_from_slice(&v.to_le_bytes());
-        self.offset += 2;
-    }
-
-    #[inline(always)]
-    pub fn put32(&mut self, v: u32) {
-        if self.offset + 4 > self.buffer.len() {
-            self.overflow = true;
-            return;
-        }
-        self.buffer[self.offset..self.offset + 4].copy_from_slice(&v.to_le_bytes());
-        self.offset += 4;
-    }
-
-    #[inline(always)]
-    pub fn put64(&mut self, v: u64) {
-        if self.offset + 8 > self.buffer.len() {
-            self.overflow = true;
-            return;
-        }
-        self.buffer[self.offset..self.offset + 8].copy_from_slice(&v.to_le_bytes());
-        self.offset += 8;
-    }
-
-    // patch a length word already written: the nested-body framing writes a
-    // placeholder, fills the body, then comes back for the count.
-    #[inline(always)]
-    pub fn patch32(&mut self, at: usize, v: u32) {
-        if at + 4 > self.buffer.len() {
-            self.overflow = true;
-            return;
-        }
-        self.buffer[at..at + 4].copy_from_slice(&v.to_le_bytes());
-    }
-}
-
-// TableReader reads the wire out of the caller's slice. A nested body is
-// read through a SUB-READER sliced out of this one, so an inner decode can
-// never reach past its own framing.
-//
-// THE REPORT IS NOT IN HERE, and the deviation from the C++ reference is
-// deliberate: a reader holding a mutable borrow of the report could not hand
-// out a sub-reader over its own buffer while that borrow stood. The report
-// is the codecs' own parameter, threaded down beside the reader — one
-// report, one caller, the same five counters.
-#[derive(Clone, Copy)]
-pub struct TableReader<'a> {
-    pub buffer: &'a [u8],
-    pub offset: usize,
-}
-
-impl<'a> TableReader<'a> {
-    #[inline(always)]
-    pub fn new(buffer: &'a [u8]) -> TableReader<'a> {
-        TableReader { buffer, offset: 0 }
-    }
-
-    #[inline(always)]
-    pub fn has(&self, bytes: u64) -> bool {
-        self.offset as u64 + bytes <= self.buffer.len() as u64
-    }
-
-    #[inline(always)]
-    pub fn get8(&mut self) -> u8 {
-        let v = self.buffer[self.offset];
-        self.offset += 1;
-        v
-    }
-
-    // The multi-byte reads take ONE range check and one load, not one check
-    // per byte: first_chunk is the safe spelling of "the next N bytes as an
-    // array", and a checked index per byte is what the generated codec's
-    // has() has already answered. The fallback is unreachable in generated
-    // code — every get is preceded by its has — and it is a zero rather than
-    // a panic because a reader that met framing damage should stop on the
-    // report, not on an abort.
-    #[inline(always)]
-    pub fn get16(&mut self) -> u16 {
-        let v = match self.buffer[self.offset..].first_chunk::<2>() {
-            Some(bytes) => u16::from_le_bytes(*bytes),
-            None => 0,
-        };
-        self.offset += 2;
-        v
-    }
-
-    #[inline(always)]
-    pub fn get32(&mut self) -> u32 {
-        let v = match self.buffer[self.offset..].first_chunk::<4>() {
-            Some(bytes) => u32::from_le_bytes(*bytes),
-            None => 0,
-        };
-        self.offset += 4;
-        v
-    }
-
-    #[inline(always)]
-    pub fn get64(&mut self) -> u64 {
-        let v = match self.buffer[self.offset..].first_chunk::<8>() {
-            Some(bytes) => u64::from_le_bytes(*bytes),
-            None => 0,
-        };
-        self.offset += 8;
-        v
-    }
-
-    // the rest of this body, as its own reader: the sub-reader a nested
-    // decode runs in.
-    #[inline(always)]
-    pub fn sub(&self, bytes: usize) -> TableReader<'a> {
-        TableReader::new(&self.buffer[self.offset..self.offset + bytes])
-    }
-
-    // skip one payload by kind; false = framing damage
-    pub fn skip(&mut self, kind: u8) -> bool {
-        match kind {
-            1 | 2 | 6 => {
-                if !self.has(1) {
-                    return false;
-                }
-                self.offset += 1;
-                true
-            }
-            3 | 7 => {
-                if !self.has(2) {
-                    return false;
-                }
-                self.offset += 2;
-                true
-            }
-            // 17 is a NODE INDEX (docs/SPEC-TABLES.md §3.1): four bytes, so it costs
-            // one row here and a reader without the kind still skips a pointer field
-            4 | 8 | 10 | 17 => {
-                if !self.has(4) {
-                    return false;
-                }
-                self.offset += 4;
-                true
-            }
-            5 | 9 | 11 => {
-                if !self.has(8) {
-                    return false;
-                }
-                self.offset += 8;
-                true
-            }
-            12 | 13 | 14 | 16 => {
-                if !self.has(4) {
-                    return false;
-                }
-                let n = self.get32() as u64;
-                if !self.has(n) {
-                    return false;
-                }
-                self.offset += n as usize;
-                true
-            }
-            // union: u16 arm id, then the arm length-prefixed (id 0 = empty, no body)
-            15 => {
-                if !self.has(2) {
-                    return false;
-                }
-                if self.get16() == 0 {
-                    return true;
-                }
-                if !self.has(4) {
-                    return false;
-                }
-                let n = self.get32() as u64;
-                if !self.has(n) {
-                    return false;
-                }
-                self.offset += n as usize;
-                true
-            }
-            _ => false,
-        }
-    }
-}
-
-// ---- reflection (tables only, docs/SPEC-TABLES.md §8) ----
-//
-// Static field descriptors for every type in the table closure: name, wire
-// id and kind, bounds, ranges, the enum/union vocabulary and its wire ids,
-// branch guards, and — because a Rust record has real offsets — where the
-// storage sits. <name>_table_type() returns the descriptor.
-
-// One arm of a union field. The payload's ADDRESS is reached through the
-// union's own accessors below rather than through an offset: Rust spells a
-// union as a real enum (SPEC §6.1's union row), which has no committed
-// payload offset for a walker to add.
-pub struct TableUnionArmInfo {
-    pub name: &'static str,
-    pub table: Option<fn() -> &'static TableTypeInfo>,
-}
-
-// A union field's tag and arms. The three function columns are what a Rust
-// enum needs where C++ needs a tag offset and an arm offset.
-pub struct TableUnionInfo {
-    // the tag ordinal a union value currently holds: 0 is None.
-    pub read_tag: fn(*const u8) -> u64,
-    // put the value at None.
-    pub clear: fn(*mut u8),
-    // select an arm, establishing the arm's declared defaults, and answer the
-    // payload's address.
-    pub select: fn(*mut u8, u64) -> *mut u8,
-    // the CURRENT arm's payload address; the tag must be the value's own.
-    pub payload: fn(*const u8, u64) -> *const u8,
-    // indexed by tag: entry 0 is None and names no table.
-    pub arms: &'static [TableUnionArmInfo],
-}
-
-// THE SHARED EMPTY DOC (docs/SPEC-TABLES.md §8.1): a declaration with no ///
-// block carries a doc column naming this one item, so absence costs a unit no
-// string data and a printer concatenates doc columns with no None test. One
-// definition for the whole unit: a STATIC, so every absent doc column is the
-// same object and comparing two by address answers the shared-string claim.
-// A const would be a value copied into each row, which is the very thing the
-// claim rules out.
-pub static TABLE_DOC_NONE: &str = "";
-
-pub struct TableFieldInfo {
-    pub name: &'static str,      // schema field name, e.g. "health"
-    pub json: &'static str,      // the TEXT form's key: json = "key", else name (§16.3)
-    pub type_name: &'static str, // schema type name, e.g. "float32", "Grade"
-    pub id: u16,                 // table-wire field id (the was alias's hash after a rename)
-    pub kind: u8,                // table-wire kind; for arrays/strings/bytes, the ELEMENT kind
-    pub is_array: bool,          // fixed or counted array (bytes included)
-    pub counted: bool,           // a _count/_length i32 companion exists
-    pub optional: bool,          // a ?T field: a _present bool companion decides whether it rides
-    pub array_bound: i32,        // array capacity / string max length; 0 for plain scalars
-    pub offset: u32,             // the storage member's offset
-    pub elem_size: u32,          // one element's size; the field's own otherwise
-    pub count_offset: u32,       // the _count/_length companion's offset, or u32::MAX
-    pub present_offset: u32,     // the _present companion's offset, or u32::MAX
-    pub table: Option<fn() -> &'static TableTypeInfo>, // nested table's descriptor
-    pub has_range: bool,         // a declared [min, max] (int or float)
-    pub range_min: f64,          // NOTE: int64 ranges beyond 2^53 lose precision here
-    pub range_max: f64,
-    // enums: the highest valid value (None = 0 is always valid); unions: the
-    // arm count (tag range [0, enum_max]); flags: the highest declared BIT
-    // INDEX; -1 otherwise.
-    pub enum_max: i64,
-    // an ENUM's or a UNION's vocabulary: value -> name, tag -> arm name.
-    pub enum_name: Option<fn(u64) -> &'static str>,
-    // a FLAGS field's vocabulary: bit index -> variant name. Rust has no
-    // overloading, so the two vocabularies are two columns where C++ has one
-    // and tells them apart by the id column being null.
-    pub flag_name: Option<fn(u32) -> &'static str>,
-    // the TABLE-WIRE id of one variant (docs/SPEC-TABLES.md §5): for an enum, the
-    // hash of the variant's name; for a union, the hash of the arm's name.
-    // 0 is the reserved id. None for every other kind.
-    pub variant_id: Option<fn(u64) -> u16>,
-    // an ENUM-KEYED array (§2.4): one slot per variant of key_type_name.
-    // Walk [0, array_bound) to print slots by name.
-    pub key_type_name: Option<&'static str>,
-    pub key_name: Option<fn(u64) -> &'static str>,
-    pub key_id: Option<fn(u64) -> u16>,
-    // union fields: the tag and its arms, behind a function so the whole
-    // descriptor stays a constant.
-    pub arms: Option<fn() -> &'static TableUnionInfo>,
-    pub guard: &'static str, // branch guard, e.g. "at_rest" or "!at_rest"; "" if unguarded
-    // what a PERSON wrote about the field (docs/SPEC-TABLES.md §8.1): the ///
-    // block above it, verbatim (SPEC §4.1), TABLE_DOC_NONE when there is
-    // none, never None, and its tags (SPEC §4.2) in declared order. An
-    // untagged field is 0 and None, the same absence every other absent
-    // column in a row spells. Static, constant data, allocating nothing.
-    pub doc: &'static str,
-    pub num_tags: i32,
-    pub tags: Option<&'static [&'static str]>,
-}
-
-pub struct TableTypeInfo {
-    pub name: &'static str, // schema type name
-    pub size: u32,          // size_of the storage struct
-    pub num_fields: i32,
-    pub fields: &'static [TableFieldInfo],
-    // put one instance back at its declared defaults, in place. A generic
-    // walker that fills a value has to be able to establish the defaults an
-    // absent field takes, and it holds no type to spell.
-    pub reset: fn(*mut u8),
-    // the declaration's own doc and tags, on the same terms as a field's
-    // (docs/SPEC-TABLES.md §8.1)
-    pub doc: &'static str,
-    pub num_tags: i32,
-    pub tags: Option<&'static [&'static str]>,
-}
-
-// ---- json walk: begin ----
-//
-// The TEXT form (docs/SPEC-TABLES.md §16): one table, one text, one walk over the
-// reflection descriptors (§8). Reading fills ONE caller-owned instance and
-// allocates nothing beyond it; writing targets a caller buffer with the
-// wire's measure/write symmetry. Everything AROUND this — which file goes
-// with which instance, what key an instance is filed under, how instances
-// link into a root table's collections — is a packer's opinion and stays
-// with the tool that holds it.
-//
-// The dialect: trailing commas are accepted on read (the authoring files
-// this exists for carry them) and never written; comments are not JSON and
-// are refused; unknown keys are skipped and counted; a duplicate key is
-// last-wins and counted; a key present with the wrong JSON type is skipped
-// and counted, never coerced.
-//
-// TWO THINGS THE C++ REFERENCE CARRIES AND THIS DOES NOT, both because Rust
-// removes the reason for them: the LOCALE's decimal point (Rust's float
-// formatting is locale-free, so a number never crosses that character), and
-// the NUL a string's storage keeps past its used length (a Rust [u8; N]
-// carries a length, not a terminator).
-
-pub const TABLE_JSON_MAX_DEPTH: i32 = 128;
-
-// A key longer than this cannot name a field, so it is skipped as unknown.
-pub const TABLE_JSON_MAX_KEY: usize = 256;
-
-// The longest numeric token the walk will convert. Anything longer is a
-// value no field can hold and counts as a kind mismatch.
-pub const TABLE_JSON_MAX_NUMBER: usize = 512;
-
-// ---- storage access: the descriptors give an offset and a width, and the
-// ---- storage is the HOST's, so every load and store goes through a width
-// ---- switch rather than a copy into the low bytes of a wider word
-
-// finite: not a NaN, not an infinity. The C++ walk spells it out because it
-// keeps its runtime surface to the handful of functions it already names;
-// Rust has the predicate, and the predicate is the same three tests.
-#[inline]
-fn table_json_finite(v: f64) -> bool {
-    v.is_finite()
-}
-
-// A vocabulary entry the descriptor could not spell. The generated name
-// functions answer "???" for a value outside the declared set, and that is
-// not a name — writing it would put a spelling in the text that the reader
-// then counts as unknown, turning a refusal into a silent loss.
-#[inline]
-fn table_json_named(name: &str) -> bool {
-    name != "???"
-}
-
-/// # Safety
-/// storage must point at width readable bytes.
-unsafe fn table_json_get_raw(storage: *const u8, width: u32) -> u64 {
-    unsafe {
-        match width {
-            1 => ptr::read_unaligned(storage) as u64,
-            2 => ptr::read_unaligned(storage as *const u16) as u64,
-            4 => ptr::read_unaligned(storage as *const u32) as u64,
-            8 => ptr::read_unaligned(storage as *const u64),
-            _ => 0,
-        }
-    }
-}
-
-/// # Safety
-/// storage must point at width writable bytes.
-unsafe fn table_json_set_raw(storage: *mut u8, width: u32, value: u64) {
-    unsafe {
-        match width {
-            1 => ptr::write_unaligned(storage, value as u8),
-            2 => ptr::write_unaligned(storage as *mut u16, value as u16),
-            4 => ptr::write_unaligned(storage as *mut u32, value as u32),
-            8 => ptr::write_unaligned(storage as *mut u64, value),
-            _ => {}
-        }
-    }
-}
-
-/// # Safety
-/// storage must point at width readable bytes.
-unsafe fn table_json_get_signed(storage: *const u8, width: u32) -> i64 {
-    let raw = unsafe { table_json_get_raw(storage, width) };
-    if width < 8 && width > 0 {
-        let sign = 1u64 << (width * 8 - 1);
-        if raw & sign != 0 {
-            return (raw | !((sign << 1) - 1)) as i64;
-        }
-    }
-    raw as i64
-}
-
-// a counted field's companion: a string's length, a bytes' length, a counted
-// array's count. Bounded by the declared extent on the way out, so a storage
-// invariant a caller broke cannot walk off the end of the array.
-/// # Safety
-/// base must point at the record this descriptor describes.
-unsafe fn table_json_count(base: *const u8, f: &TableFieldInfo) -> i32 {
-    if !f.counted {
-        return f.array_bound;
-    }
-    let mut count = unsafe { ptr::read_unaligned(base.add(f.count_offset as usize) as *const i32) };
-    if count < 0 {
-        count = 0;
-    }
-    if count > f.array_bound {
-        count = f.array_bound;
-    }
-    count
-}
-
-/// # Safety
-/// base must point at the record this descriptor describes.
-unsafe fn table_json_set_count(base: *mut u8, f: &TableFieldInfo, count: i32) {
-    if f.counted {
-        unsafe { ptr::write_unaligned(base.add(f.count_offset as usize) as *mut i32, count) };
-    }
-}
-
-// ---- what a field's kind expects to see in the text ----
-//
-// One classifier, consulted by both directions, so a reader and a writer can
-// never disagree about a kind's JSON form. 'o' object, 'a' array, 's'
-// string, 'n' number, 'b' boolean.
-//
-// bytes(N) is the one kind whose element kind does not decide its form: it
-// shares u8 with a plain array of u8, and rides as base64. The schema type
-// name settles it, and "bytes" is a keyword no declaration can claim.
-#[inline]
-fn table_json_is_bytes(f: &TableFieldInfo) -> bool {
-    f.is_array && f.kind == 6 && f.type_name == "bytes"
-}
-
-// An ENUM-KEYED array (docs/SPEC-TABLES.md §2.4): its JSON form is an OBJECT
-// keyed by variant name, not a positional array, because that is what the
-// storage is — one slot per variant, addressed by the variant.
-#[inline]
-fn table_json_is_keyed(f: &TableFieldInfo) -> bool {
-    f.key_name.is_some()
-}
-
-// THE KEY A STORAGE SLOT HOLDS (§2.4, §8): the storage shifts left, so slot
-// i holds the key i + 1 and nothing is stored for None.
-#[inline]
-fn table_json_keyed_slot_key(slot: i64) -> u64 {
-    slot as u64 + 1
-}
-
-// A slot whose key names a variant of the keying enum. Every slot in
-// [0, array_bound) does, unless the enum carries max-headroom variants
-// outside a table closure, where a reserved value names nothing and its key
-// id is 0 — the reserved id no declared name can fold to (§5).
-#[inline]
-fn table_json_keyed_slot_valid(f: &TableFieldInfo, slot: i64) -> bool {
-    match f.key_id {
-        Some(key_id) => key_id(table_json_keyed_slot_key(slot)) != 0,
-        None => false,
-    }
-}
-
-#[inline]
-fn table_json_is_flags(f: &TableFieldInfo) -> bool {
-    f.flag_name.is_some()
-}
-
-#[inline]
-fn table_json_is_enum(f: &TableFieldInfo) -> bool {
-    f.variant_id.is_some() && f.arms.is_none()
-}
-
-fn table_json_shape(f: &TableFieldInfo) -> u8 {
-    if f.kind == 12 {
-        return b's'; // string
-    }
-    if table_json_is_bytes(f) {
-        return b's'; // bytes: base64
-    }
-    if table_json_is_keyed(f) {
-        return b'o'; // an object keyed by variant NAME
-    }
-    if f.is_array {
-        return b'a';
-    }
-    if f.arms.is_some() {
-        return b'o'; // union: an object with ONE key
-    }
-    if f.kind == 13 {
-        return b'o'; // nested table or type
-    }
-    if table_json_is_enum(f) {
-        return b's';
-    }
-    if table_json_is_flags(f) {
-        return b'a';
-    }
-    if f.kind == 1 {
-        return b'b';
-    }
-    b'n'
-}
-
-// the ELEMENT shape of an array field — the same classifier one level down
-fn table_json_element_shape(f: &TableFieldInfo) -> u8 {
-    if f.kind == 13 {
-        return b'o';
-    }
-    if table_json_is_enum(f) {
-        return b's';
-    }
-    if table_json_is_flags(f) {
-        return b'a';
-    }
-    if f.kind == 1 {
-        return b'b';
-    }
-    b'n'
-}
-
-// A guarded group rides only when its guard reads true — the wire's own
-// elision (§4), carried into the text so a text and a wire written from one
-// instance say the same thing.
-/// # Safety
-/// base must point at an instance of the record info describes.
-unsafe fn table_json_guard_holds(base: *const u8, info: &TableTypeInfo, guard: &str) -> bool {
-    let bytes = guard.as_bytes();
-    let mut p = 0usize;
-    loop {
-        while p < bytes.len() && (bytes[p] == b' ' || bytes[p] == b'&') {
-            p += 1;
-        }
-        if p >= bytes.len() {
-            return true;
-        }
-        let mut want = true;
-        if bytes[p] == b'!' {
-            want = false;
-            p += 1;
-        }
-        let start = p;
-        while p < bytes.len() && bytes[p] != b' ' && bytes[p] != b'&' {
-            p += 1;
-        }
-        let name = &guard[start..p];
-        let mut value = false;
-        for f in info.fields.iter() {
-            if f.name == name {
-                value = unsafe { table_json_get_raw(base.add(f.offset as usize), f.elem_size) } != 0;
-                break;
-            }
-        }
-        if value != want {
-            return false;
-        }
-    }
-}
-
-// ---- writing ----
-
-// The writer sink MEASURES when it holds no buffer and WRITES when it does,
-// over one code path — so measure and write agree byte for byte, the wire's
-// invariant (§9) carried across.
-struct TableJsonOut<'a> {
-    buffer: Option<&'a mut [u8]>,
-    offset: i64,
-    overflow: bool,
-}
-
-impl TableJsonOut<'_> {
-    fn raw(&mut self, data: &[u8]) {
-        if let Some(buffer) = self.buffer.as_deref_mut() {
-            if self.offset as usize + data.len() > buffer.len() {
-                self.overflow = true;
-                return;
-            }
-            buffer[self.offset as usize..self.offset as usize + data.len()].copy_from_slice(data);
-        }
-        self.offset += data.len() as i64;
-    }
-
-    fn put(&mut self, c: u8) {
-        self.raw(&[c]);
-    }
-
-    fn text(&mut self, s: &str) {
-        self.raw(s.as_bytes());
-    }
-
-    fn line(&mut self, depth: i32) {
-        self.put(b'\n');
-        for _ in 0..depth {
-            self.raw(b"  ");
-        }
-    }
-}
-
-const TABLE_JSON_BASE64: &[u8; 64] =
-    b"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
-
-// Byte-indexed sextets; -1 preserves invalid-character handling.
-const TABLE_JSON_BASE64_DECODE: [i8; 256] = [
-    -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-    -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-    -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 62, -1, -1, -1, 63,
-    52, 53, 54, 55, 56, 57, 58, 59, 60, 61, -1, -1, -1, -1, -1, -1,
-    -1,  0,  1,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 12, 13, 14,
-    15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, -1, -1, -1, -1, -1,
-    -1, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40,
-    41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, -1, -1, -1, -1, -1,
-    -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-    -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-    -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-    -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-    -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-    -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-    -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-    -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-];
-
-fn table_json_write_base64(out: &mut TableJsonOut, data: &[u8]) {
-    let alphabet = TABLE_JSON_BASE64;
-    out.put(b'"');
-    let length = data.len();
-    let mut i = 0usize;
-    while i + 3 <= length {
-        let triple =
-            ((data[i] as u32) << 16) | ((data[i + 1] as u32) << 8) | (data[i + 2] as u32);
-        let quad = [
-            alphabet[((triple >> 18) & 0x3f) as usize],
-            alphabet[((triple >> 12) & 0x3f) as usize],
-            alphabet[((triple >> 6) & 0x3f) as usize],
-            alphabet[(triple & 0x3f) as usize],
-        ];
-        out.raw(&quad);
-        i += 3;
-    }
-    if i < length {
-        let left = length - i;
-        let mut triple = (data[i] as u32) << 16;
-        if left == 2 {
-            triple |= (data[i + 1] as u32) << 8;
-        }
-        let mut quad = [
-            alphabet[((triple >> 18) & 0x3f) as usize],
-            alphabet[((triple >> 12) & 0x3f) as usize],
-            b'=',
-            b'=',
-        ];
-        if left == 2 {
-            quad[2] = alphabet[((triple >> 6) & 0x3f) as usize];
-        }
-        out.raw(&quad);
-    }
-    out.put(b'"');
-}
-
-// One UTF-8 sequence at s, or None when the bytes there are not one. Rejects
-// the lot: a stray continuation, an overlong form, a surrogate half, and
-// anything past U+10FFFF.
-fn table_json_utf8(s: &[u8]) -> Option<usize> {
-    let lead = s[0];
-    let (want, mut code) = if lead < 0x80 {
-        return Some(1);
-    } else if (0xc2..=0xdf).contains(&lead) {
-        (2usize, (lead & 0x1f) as u32)
-    } else if (0xe0..=0xef).contains(&lead) {
-        (3usize, (lead & 0x0f) as u32)
-    } else if (0xf0..=0xf4).contains(&lead) {
-        (4usize, (lead & 0x07) as u32)
-    } else {
-        return None;
-    };
-    if s.len() < want {
-        return None;
-    }
-    for i in 1..want {
-        let next = s[i];
-        if next & 0xc0 != 0x80 {
-            return None;
-        }
-        code = (code << 6) | (next & 0x3f) as u32;
-    }
-    if want == 3 && code < 0x800 {
-        return None; // overlong
-    }
-    if want == 4 && code < 0x10000 {
-        return None; // overlong
-    }
-    if (0xd800..=0xdfff).contains(&code) {
-        return None; // a surrogate half
-    }
-    if code > 0x10ffff {
-        return None;
-    }
-    Some(want)
-}
-
-// A JSON text MUST be valid UTF-8 (RFC 8259 §8.1). The read path is
-// byte-transparent — the wire imposes no encoding (§3) and a string may hold
-// anything — so the WRITER is where that obligation is met: a byte that is
-// not part of a well-formed sequence is written as U+FFFD, one per bad byte,
-// and never raw.
-fn table_json_write_string_bytes(out: &mut TableJsonOut, s: &[u8]) {
-    const HEX: &[u8; 16] = b"0123456789abcdef";
-    out.put(b'"');
-    let mut i = 0usize;
-    while i < s.len() {
-        let c = s[i];
-        match c {
-            b'"' => out.raw(b"\\\""),
-            b'\\' => out.raw(b"\\\\"),
-            0x08 => out.raw(b"\\b"),
-            0x0c => out.raw(b"\\f"),
-            b'\n' => out.raw(b"\\n"),
-            b'\r' => out.raw(b"\\r"),
-            b'\t' => out.raw(b"\\t"),
-            _ => {
-                if c < 0x20 {
-                    let escape = [
-                        b'\\',
-                        b'u',
-                        b'0',
-                        b'0',
-                        HEX[(c >> 4) as usize],
-                        HEX[(c & 0xf) as usize],
-                    ];
-                    out.raw(&escape);
-                } else if c < 0x80 {
-                    out.put(c);
-                } else {
-                    match table_json_utf8(&s[i..]) {
-                        None => out.raw(&[0xef, 0xbf, 0xbd]), // U+FFFD, one per bad byte
-                        Some(width) => {
-                            out.raw(&s[i..i + width]);
-                            i += width - 1;
-                        }
-                    }
-                }
-            }
-        }
-        i += 1;
-    }
-    out.put(b'"');
-}
-
-#[inline]
-fn table_json_write_string(out: &mut TableJsonOut, s: &str) {
-    table_json_write_string_bytes(out, s.as_bytes());
-}
-
-fn table_json_write_unsigned(out: &mut TableJsonOut, mut value: u64) {
-    let mut digits = [0u8; 24];
-    let mut n = 0usize;
-    loop {
-        digits[n] = b'0' + (value % 10) as u8;
-        n += 1;
-        value /= 10;
-        if value == 0 {
-            break;
-        }
-    }
-    let mut text = [0u8; 24];
-    for i in 0..n {
-        text[i] = digits[n - 1 - i];
-    }
-    out.raw(&text[..n]);
-}
-
-fn table_json_write_signed(out: &mut TableJsonOut, value: i64) {
-    if value < 0 {
-        out.put(b'-');
-        table_json_write_unsigned(out, (value as u64).wrapping_neg());
-        return;
-    }
-    table_json_write_unsigned(out, value as u64);
-}
-
-// C's "%.*g", exactly, because the goldens are its bytes: convert to
-// precision significant digits, choose style e when the decimal exponent is
-// below -4 or at least the precision and style f otherwise, then strip the
-// fraction's trailing zeros and a bare decimal point. The exponent field
-// carries a sign and at least two digits.
-// THE NUMBER SINK, and it is a STACK buffer because docs/USAGE.md's promise is
-// that the text path allocates nothing beyond the instance and the buffers the
-// caller passed. core::fmt writes a float through its own stack buffers, so a
-// sink that does not allocate leaves the whole formatter allocation-free, and
-// make tables-rust-alloc-audit is what counts it.
-//
-// SIXTY-FOUR BYTES, the C++ walker's own char text[64], and it is provably
-// enough rather than generously chosen: style f is used only when the decimal
-// exponent is in [-4, precision), so the integer part is at most 18 digits and
-// the fraction at most 20, and style e is a mantissa of at most 18 plus four
-// characters of exponent. A write past it REFUSES, exactly as the C++ walker
-// refuses a snprintf that would not fit — a surprise is a refusal here, never
-// a truncated number.
-const TABLE_JSON_DIGITS: usize = 64;
-
-struct TableJsonDigits {
-    bytes: [u8; TABLE_JSON_DIGITS],
-    length: usize,
-}
-
-impl TableJsonDigits {
-    fn new() -> TableJsonDigits {
-        TableJsonDigits {
-            bytes: [0; TABLE_JSON_DIGITS],
-            length: 0,
-        }
-    }
-
-    fn clear(&mut self) {
-        self.length = 0;
-    }
-
-    fn as_str(&self) -> &str {
-        core::str::from_utf8(&self.bytes[..self.length]).unwrap_or("")
-    }
-
-    fn as_bytes(&self) -> &[u8] {
-        &self.bytes[..self.length]
-    }
-
-    fn push(&mut self, b: u8) -> bool {
-        if self.length >= self.bytes.len() {
-            return false;
-        }
-        self.bytes[self.length] = b;
-        self.length += 1;
-        true
-    }
-
-    // strip the fraction's trailing zeros and then a bare decimal point, which
-    // is what %g does once it has converted
-    fn trim_fraction(&mut self) {
-        if !self.as_bytes().contains(&b'.') {
-            return;
-        }
-        while self.length > 0 && self.bytes[self.length - 1] == b'0' {
-            self.length -= 1;
-        }
-        if self.length > 0 && self.bytes[self.length - 1] == b'.' {
-            self.length -= 1;
-        }
-    }
-}
-
-impl core::fmt::Write for TableJsonDigits {
-    fn write_str(&mut self, s: &str) -> core::fmt::Result {
-        let end = self.length + s.len();
-        if end > self.bytes.len() {
-            return Err(core::fmt::Error);
-        }
-        self.bytes[self.length..end].copy_from_slice(s.as_bytes());
-        self.length = end;
-        Ok(())
-    }
-}
-
-// C's "%.*g", exactly, because the goldens are its bytes: convert to
-// precision significant digits, choose style e when the decimal exponent is
-// below -4 or at least the precision and style f otherwise, then strip the
-// fraction's trailing zeros and a bare decimal point. The exponent field
-// carries a sign and at least two digits.
-fn table_json_format_g(
-    out: &mut TableJsonDigits,
-    scratch: &mut TableJsonDigits,
-    value: f64,
-    precision: usize,
-) -> bool {
-    use core::fmt::Write;
-    let precision = if precision == 0 { 1 } else { precision };
-    // {:e} gives one digit before the point and the exponent of the value
-    // ROUNDED to precision significant digits, which is the exponent C's %g
-    // makes its style decision on.
-    scratch.clear();
-    if write!(scratch, "{:.*e}", precision - 1, value).is_err() {
-        return false;
-    }
-    let scientific = scratch.as_str();
-    let (mantissa, exponent) = match scientific.split_once('e') {
-        Some((m, e)) => (m, e.parse::<i32>().unwrap_or(0)),
-        None => (scientific, 0),
-    };
-    out.clear();
-    if exponent < -4 || exponent >= precision as i32 {
-        for b in mantissa.as_bytes() {
-            if !out.push(*b) {
-                return false;
-            }
-        }
-        out.trim_fraction();
-        if !out.push(b'e') || !out.push(if exponent < 0 { b'-' } else { b'+' }) {
-            return false;
-        }
-        return write!(out, "{:02}", exponent.unsigned_abs()).is_ok();
-    }
-    let places = (precision as i32 - 1 - exponent).max(0) as usize;
-    if write!(out, "{:.*}", places, value).is_err() {
-        return false;
-    }
-    out.trim_fraction();
-    true
-}
-
-// A float writes at the SHORTEST precision that reads back as the same value
-// at the field's own width, so a round trip is exact and a text stays
-// readable. Non-finite values have no JSON spelling at all, and the writer
-// REFUSES rather than losing one silently — the same rule measure and save
-// already apply to an enum value no variant names (§5).
-fn table_json_write_float(out: &mut TableJsonOut, value: f64, single: bool) -> bool {
-    if !table_json_finite(value) {
-        return false;
-    }
-    let low = if single { 6 } else { 15 };
-    let high = if single { 9 } else { 17 };
-    let mut text = TableJsonDigits::new();
-    let mut scratch = TableJsonDigits::new();
-    let mut digits = low;
-    loop {
-        if !table_json_format_g(&mut text, &mut scratch, value, digits) {
-            return false;
-        }
-        if digits >= high {
-            break;
-        }
-        if single {
-            if text.as_str().parse::<f32>().map(|v| v as f64) == Ok(value) {
-                break;
-            }
-        } else if text.as_str().parse::<f64>() == Ok(value) {
-            break;
-        }
-        digits += 1;
-    }
-    out.raw(text.as_bytes());
-    true
-}
-
-// one scalar, at one storage address: a nested object, a union, a
-// vocabulary, or a number
-/// # Safety
-/// storage must point at the field's own slot.
-unsafe fn table_json_write_scalar(
-    out: &mut TableJsonOut,
-    storage: *const u8,
-    f: &TableFieldInfo,
-    depth: i32,
-) -> bool {
-    unsafe {
-        if let Some(arms_of) = f.arms {
-            // a union is an object with ONE key, the arm's name; None is {}
-            let arms = arms_of();
-            let tag = (arms.read_tag)(storage);
-            if tag == 0 {
-                out.raw(b"{}");
-                return true;
-            }
-            if tag as i64 > f.enum_max {
-                return false; // a tag no arm names, exactly as measure refuses it
-            }
-            let name_of = match f.enum_name {
-                Some(n) => n,
-                None => return false,
-            };
-            let arm = name_of(tag);
-            // and refuse on the NAME, not merely on the bound (§16.2)
-            if !table_json_named(arm) {
-                return false;
-            }
-            let entry = &arms.arms[tag as usize];
-            let table = match entry.table {
-                Some(t) => t,
-                None => return false,
-            };
-            out.put(b'{');
-            out.line(depth + 1);
-            table_json_write_string(out, arm);
-            out.raw(b": ");
-            if !table_json_write_value(out, (arms.payload)(storage, tag), table(), depth + 1) {
-                return false;
-            }
-            out.line(depth);
-            out.put(b'}');
-            return true;
-        }
-        if f.kind == 13 {
-            let table = match f.table {
-                Some(t) => t,
-                None => return false,
-            };
-            return table_json_write_value(out, storage, table(), depth);
-        }
-        if table_json_is_enum(f) {
-            let value = table_json_get_raw(storage, f.elem_size);
-            // a value no variant names has no text spelling, exactly as it has
-            // no wire identity (docs/SPEC-TABLES.md §5)
-            if value as i64 > f.enum_max {
-                return false;
-            }
-            let variant_id = f.variant_id.unwrap();
-            if value != 0 && variant_id(value) == 0 {
-                return false;
-            }
-            let name = f.enum_name.unwrap()(value);
-            if !table_json_named(name) {
-                return false;
-            }
-            table_json_write_string(out, name);
-            return true;
-        }
-        if table_json_is_flags(f) {
-            let bits = table_json_get_raw(storage, f.elem_size);
-            if bits == 0 {
-                out.raw(b"[]");
-                return true;
-            }
-            let flag_name = f.flag_name.unwrap();
-            out.put(b'[');
-            let mut first = true;
-            for bit in 0..64i64 {
-                if bits & (1u64 << bit) == 0 {
-                    continue;
-                }
-                if bit > f.enum_max {
-                    return false; // a bit no variant names has no text spelling
-                }
-                let name = flag_name(bit as u32);
-                if !table_json_named(name) {
-                    return false;
-                }
-                if !first {
-                    out.put(b',');
-                }
-                first = false;
-                out.line(depth + 1);
-                table_json_write_string(out, name);
-            }
-            out.line(depth);
-            out.put(b']');
-            return true;
-        }
-        match f.kind {
-            1 => {
-                out.text(if table_json_get_raw(storage, f.elem_size) != 0 {
-                    "true"
-                } else {
-                    "false"
-                });
-                true
-            }
-            10 => {
-                let v = f32::from_bits(table_json_get_raw(storage, 4) as u32);
-                table_json_write_float(out, v as f64, true)
-            }
-            11 => {
-                let v = f64::from_bits(table_json_get_raw(storage, 8));
-                table_json_write_float(out, v, false)
-            }
-            2..=5 => {
-                table_json_write_signed(out, table_json_get_signed(storage, f.elem_size));
-                true
-            }
-            _ => {
-                table_json_write_unsigned(out, table_json_get_raw(storage, f.elem_size));
-                true
-            }
-        }
-    }
-}
-
-/// # Safety
-/// base must point at an instance of the record the descriptor describes.
-unsafe fn table_json_write_field(
-    out: &mut TableJsonOut,
-    base: *const u8,
-    f: &TableFieldInfo,
-    depth: i32,
-) -> bool {
-    unsafe {
-        let storage = base.add(f.offset as usize);
-        if f.kind == 12 {
-            let used = table_json_count(base, f) as usize;
-            table_json_write_string_bytes(out, core::slice::from_raw_parts(storage, used));
-            return true;
-        }
-        if table_json_is_bytes(f) {
-            let used = table_json_count(base, f) as usize;
-            table_json_write_base64(out, core::slice::from_raw_parts(storage, used));
-            return true;
-        }
-        if table_json_is_keyed(f) {
-            // one entry per SLOT, keyed by the variant that owns it, so
-            // inserting a variant next season moves nothing in the text
-            // either. Slot i holds the key i + 1: nothing is stored for None,
-            // so nothing is written for it.
-            out.put(b'{');
-            let mut first = true;
-            for slot in 0..f.array_bound as i64 {
-                if !table_json_keyed_slot_valid(f, slot) {
-                    continue;
-                }
-                if !first {
-                    out.put(b',');
-                }
-                first = false;
-                out.line(depth + 1);
-                let key = f.key_name.unwrap()(table_json_keyed_slot_key(slot));
-                table_json_write_string(out, key);
-                out.raw(b": ");
-                if !table_json_write_scalar(
-                    out,
-                    storage.add(slot as usize * f.elem_size as usize),
-                    f,
-                    depth + 1,
-                ) {
-                    return false;
-                }
-            }
-            if first {
-                out.raw(b"}");
-                return true;
-            }
-            out.line(depth);
-            out.put(b'}');
-            return true;
-        }
-        if f.is_array {
-            let count = table_json_count(base, f);
-            if count == 0 {
-                out.raw(b"[]");
-                return true;
-            }
-            out.put(b'[');
-            for i in 0..count {
-                if i > 0 {
-                    out.put(b',');
-                }
-                out.line(depth + 1);
-                if !table_json_write_scalar(
-                    out,
-                    storage.add(i as usize * f.elem_size as usize),
-                    f,
-                    depth + 1,
-                ) {
-                    return false;
-                }
-            }
-            out.line(depth);
-            out.put(b']');
-            return true;
-        }
-        table_json_write_scalar(out, storage, f, depth)
-    }
-}
-
-// One instance, every field, in DECLARATION ORDER, defaults included — a
-// text is for people and tools, and a text that elides is a text a reader
-// has to know the schema to complete.
-/// # Safety
-/// base must point at an instance of the record info describes.
-unsafe fn table_json_write_value(
-    out: &mut TableJsonOut,
-    base: *const u8,
-    info: &TableTypeInfo,
-    depth: i32,
-) -> bool {
-    unsafe {
-        let mut any = false;
-        for f in info.fields.iter() {
-            if !f.guard.is_empty() && !table_json_guard_holds(base, info, f.guard) {
-                continue;
-            }
-            // an ABSENT optional writes no key: presence of the key IS the
-            // presence (§16.2), so an absent field is an absent key and
-            // nothing else would read back as absent
-            if f.optional && table_json_get_raw(base.add(f.present_offset as usize), 1) == 0 {
-                continue;
-            }
-            if !any {
-                out.put(b'{');
-            } else {
-                out.put(b',');
-            }
-            any = true;
-            out.line(depth + 1);
-            table_json_write_string(out, f.json);
-            out.raw(b": ");
-            if !table_json_write_field(out, base, f, depth + 1) {
-                return false;
-            }
-        }
-        if !any {
-            out.raw(b"{}");
-            return true;
-        }
-        out.line(depth);
-        out.put(b'}');
-        true
-    }
-}
-
-// ---- reading ----
-
-struct TableJsonIn<'a> {
-    text: &'a [u8],
-    pos: usize,
-    bad: bool, // the text is not JSON: the walk stops and keeps what it placed
-}
-
-impl TableJsonIn<'_> {
-    fn space(&mut self) {
-        while self.pos < self.text.len() {
-            let c = self.text[self.pos];
-            if c == b' ' || c == b'\t' || c == b'\n' || c == b'\r' {
-                self.pos += 1;
-                continue;
-            }
-            // comments are not JSON, and a walk that guessed at one would be
-            // reading a dialect nobody wrote down
-            if c == b'/' {
-                self.bad = true;
-            }
-            return;
-        }
-    }
-
-    fn peek(&mut self) -> u8 {
-        self.space();
-        if self.pos < self.text.len() {
-            self.text[self.pos]
-        } else {
-            0
-        }
-    }
-
-    // the shape of the value sitting at the cursor, without consuming it
-    fn value_shape(&mut self) -> u8 {
-        match self.peek() {
-            b'{' => b'o',
-            b'[' => b'a',
-            b'"' => b's',
-            b't' | b'f' => b'b',
-            b'n' => b'z',
-            0 => 0,
-            _ => b'n',
-        }
-    }
-
-    fn literal(&mut self, word: &[u8]) -> bool {
-        if self.pos + word.len() > self.text.len()
-            || &self.text[self.pos..self.pos + word.len()] != word
-        {
-            self.bad = true;
-            return false;
-        }
-        self.pos += word.len();
-        true
-    }
-
-    // one \uXXXX escape body; -1 when the four hex digits are not there
-    fn hex4(&mut self) -> i32 {
-        if self.pos + 4 > self.text.len() {
-            return -1;
-        }
-        let mut value = 0i32;
-        for i in 0..4 {
-            let c = self.text[self.pos + i];
-            let digit = match c {
-                b'0'..=b'9' => (c - b'0') as i32,
-                b'a'..=b'f' => (c - b'a') as i32 + 10,
-                b'A'..=b'F' => (c - b'A') as i32 + 10,
-                _ => return -1,
-            };
-            value = (value << 4) | digit;
-        }
-        self.pos += 4;
-        value
-    }
-}
-
-fn table_json_encode_utf8(code: u32, unit: &mut [u8; 4]) -> usize {
-    if code < 0x80 {
-        unit[0] = code as u8;
-        return 1;
-    }
-    if code < 0x800 {
-        unit[0] = 0xc0 | (code >> 6) as u8;
-        unit[1] = 0x80 | (code & 0x3f) as u8;
-        return 2;
-    }
-    if code < 0x10000 {
-        unit[0] = 0xe0 | (code >> 12) as u8;
-        unit[1] = 0x80 | ((code >> 6) & 0x3f) as u8;
-        unit[2] = 0x80 | (code & 0x3f) as u8;
-        return 3;
-    }
-    unit[0] = 0xf0 | (code >> 18) as u8;
-    unit[1] = 0x80 | ((code >> 12) & 0x3f) as u8;
-    unit[2] = 0x80 | ((code >> 6) & 0x3f) as u8;
-    unit[3] = 0x80 | (code & 0x3f) as u8;
-    4
-}
-
-// Where a scanned string goes: a raw storage slot with a capacity, or
-// nowhere at all (scanning past a string without keeping it).
-enum TableJsonSink {
-    Discard,
-    // # Safety: the pointer covers capacity writable bytes.
-    Storage(*mut u8, usize),
-}
-
-// Scan one JSON string into a caller buffer. Bytes are appended ONE CODE
-// POINT AT A TIME — an escape's encoding, or a UTF-8 sequence read whole —
-// so a string longer than the field is clamped AT A CODE POINT BOUNDARY and
-// never cut through a multi-byte character. Clamping is counted, never
-// fatal, exactly as it is on the wire (§4).
-fn table_json_scan_string(
-    input: &mut TableJsonIn,
-    sink: TableJsonSink,
-    report: &mut TableReport,
-) -> Option<i32> {
-    if input.peek() != b'"' {
-        input.bad = true;
-        return None;
-    }
-    input.pos += 1;
-    let mut placed = 0i32;
-    let mut clamped = false;
-    loop {
-        if input.pos >= input.text.len() {
-            input.bad = true;
-            return None;
-        }
-        let c = input.text[input.pos];
-        if c == b'"' {
-            input.pos += 1;
-            break;
-        }
-        let mut unit = [0u8; 4];
-        let unit_length;
-        if c == b'\\' {
-            input.pos += 1;
-            if input.pos >= input.text.len() {
-                input.bad = true;
-                return None;
-            }
-            let escape = input.text[input.pos];
-            input.pos += 1;
-            match escape {
-                b'"' => {
-                    unit[0] = b'"';
-                    unit_length = 1;
-                }
-                b'\\' => {
-                    unit[0] = b'\\';
-                    unit_length = 1;
-                }
-                b'/' => {
-                    unit[0] = b'/';
-                    unit_length = 1;
-                }
-                b'b' => {
-                    unit[0] = 0x08;
-                    unit_length = 1;
-                }
-                b'f' => {
-                    unit[0] = 0x0c;
-                    unit_length = 1;
-                }
-                b'n' => {
-                    unit[0] = b'\n';
-                    unit_length = 1;
-                }
-                b'r' => {
-                    unit[0] = b'\r';
-                    unit_length = 1;
-                }
-                b't' => {
-                    unit[0] = b'\t';
-                    unit_length = 1;
-                }
-                b'u' => {
-                    let high = input.hex4();
-                    if high < 0 {
-                        input.bad = true;
-                        return None;
-                    }
-                    let mut code = high as u32;
-                    if (0xd800..=0xdbff).contains(&high)
-                        && input.pos + 2 <= input.text.len()
-                        && input.text[input.pos] == b'\\'
-                        && input.text[input.pos + 1] == b'u'
-                    {
-                        let mark = input.pos;
-                        input.pos += 2;
-                        let low = input.hex4();
-                        if (0xdc00..=0xdfff).contains(&low) {
-                            code = 0x10000
-                                + (((high as u32) - 0xd800) << 10)
-                                + ((low as u32) - 0xdc00);
-                        } else {
-                            input.pos = mark; // a lone lead surrogate rides as itself
-                        }
-                    }
-                    // a surrogate half that never found its partner has no
-                    // UTF-8 encoding: encoding it anyway would manufacture
-                    // CESU-8 out of input that was valid JSON, so it reads as
-                    // the replacement character
-                    if (0xd800..=0xdfff).contains(&code) {
-                        code = 0xfffd;
-                    }
-                    unit_length = table_json_encode_utf8(code, &mut unit);
-                }
-                _ => {
-                    input.bad = true;
-                    return None;
-                }
-            }
-        } else if c < 0x20 {
-            input.bad = true; // a raw control character is not a JSON string body
-            return None;
-        } else {
-            // a UTF-8 sequence read WHOLE, so the clamp below can only land
-            // between code points. Only bytes that ACTUALLY look like
-            // continuations are taken: the wire imposes no encoding (§3), so a
-            // string may legitimately hold a stray lead byte, and one at the
-            // end of a text must not swallow the closing quote.
-            let want = if c & 0xe0 == 0xc0 {
-                2
-            } else if c & 0xf0 == 0xe0 {
-                3
-            } else if c & 0xf8 == 0xf0 {
-                4
-            } else {
-                1
-            };
-            unit[0] = c;
-            input.pos += 1;
-            let mut got = 1usize;
-            while got < want
-                && input.pos < input.text.len()
-                && input.text[input.pos] & 0xc0 == 0x80
-            {
-                unit[got] = input.text[input.pos];
-                got += 1;
-                input.pos += 1;
-            }
-            unit_length = got;
-        }
-        if let TableJsonSink::Storage(out, capacity) = sink {
-            if placed as usize + unit_length <= capacity {
-                unsafe {
-                    ptr::copy_nonoverlapping(unit.as_ptr(), out.add(placed as usize), unit_length)
-                };
-                placed += unit_length as i32;
-            } else {
-                clamped = true;
-            }
-        }
-    }
-    if clamped {
-        report.clamped += 1;
-    }
-    Some(placed)
-}
-
-// A scanned key, kept on the stack: a key longer than the cap cannot name a
-// field and is skipped as unknown, which is exactly what a clamped scan
-// leaves behind.
-struct TableJsonKey {
-    bytes: [u8; TABLE_JSON_MAX_KEY],
-    length: usize,
-}
-
-impl TableJsonKey {
-    fn new() -> TableJsonKey {
-        TableJsonKey {
-            bytes: [0; TABLE_JSON_MAX_KEY],
-            length: 0,
-        }
-    }
-
-    fn as_str(&self) -> &str {
-        core::str::from_utf8(&self.bytes[..self.length]).unwrap_or("")
-    }
-}
-
-fn table_json_scan_key(
-    input: &mut TableJsonIn,
-    key: &mut TableJsonKey,
-    report: &mut TableReport,
-) -> bool {
-    let sink = TableJsonSink::Storage(key.bytes.as_mut_ptr(), TABLE_JSON_MAX_KEY - 1);
-    match table_json_scan_string(input, sink, report) {
-        Some(length) => {
-            key.length = length as usize;
-            true
-        }
-        None => false,
-    }
-}
-
-// Scan one number, to JSON's OWN grammar (RFC 8259 §6) and not to a run of
-// number-ish characters:
-//
-//     number = [ "-" ] int [ frac ] [ exp ]
-//     int    = "0" / ( digit1-9 *digit )
-//     frac   = "." 1*digit
-//     exp    = ( "e" / "E" ) [ "-" / "+" ] 1*digit
-//
-// Scanning the production is what makes a typo in an authoring file a
-// DIAGNOSTIC rather than a value: "1-2" scans as 1 and leaves "-2" where the
-// object expects a comma, so the text is malformed — which is what §16.2
-// already promises.
-fn table_json_walk_number(input: &mut TableJsonIn) -> Option<bool> {
-    input.space();
-    let mut whole = true;
-    let text = input.text;
-    if input.pos < text.len() && text[input.pos] == b'-' {
-        input.pos += 1;
-    }
-    if input.pos >= text.len() {
-        return None;
-    }
-    if text[input.pos] == b'0' {
-        input.pos += 1;
-    } else if text[input.pos].is_ascii_digit() {
-        while input.pos < text.len() && text[input.pos].is_ascii_digit() {
-            input.pos += 1;
-        }
-    } else {
-        return None;
-    }
-    if input.pos < text.len() && text[input.pos] == b'.' {
-        input.pos += 1;
-        if input.pos >= text.len() || !text[input.pos].is_ascii_digit() {
-            return None;
-        }
-        while input.pos < text.len() && text[input.pos].is_ascii_digit() {
-            input.pos += 1;
-        }
-        whole = false;
-    }
-    if input.pos < text.len() && (text[input.pos] == b'e' || text[input.pos] == b'E') {
-        input.pos += 1;
-        if input.pos < text.len() && (text[input.pos] == b'-' || text[input.pos] == b'+') {
-            input.pos += 1;
-        }
-        if input.pos >= text.len() || !text[input.pos].is_ascii_digit() {
-            return None;
-        }
-        while input.pos < text.len() && text[input.pos].is_ascii_digit() {
-            input.pos += 1;
-        }
-        whole = false;
-    }
-    Some(whole)
-}
-
-// the same production, with the token kept for conversion
-struct TableJsonNumber {
-    bytes: [u8; TABLE_JSON_MAX_NUMBER],
-    length: usize,
-    integral: bool,
-}
-
-fn table_json_scan_number(input: &mut TableJsonIn) -> Option<TableJsonNumber> {
-    input.space();
-    let start = input.pos;
-    let integral = table_json_walk_number(input)?;
-    let count = input.pos - start;
-    if count == 0 || count >= TABLE_JSON_MAX_NUMBER {
-        return None;
-    }
-    let mut token = TableJsonNumber {
-        bytes: [0; TABLE_JSON_MAX_NUMBER],
-        length: count,
-        integral,
-    };
-    token.bytes[..count].copy_from_slice(&input.text[start..start + count]);
-    Some(token)
-}
-
-impl TableJsonNumber {
-    fn as_str(&self) -> &str {
-        core::str::from_utf8(&self.bytes[..self.length]).unwrap_or("0")
-    }
-
-    // the token's exact double. Rust's float parse is correctly rounded and
-    // locale-free, so no character crosses a decimal point on the way in.
-    fn to_double(&self, single: bool) -> f64 {
-        let text = self.as_str();
-        if single {
-            return text.parse::<f32>().unwrap_or(0.0) as f64;
-        }
-        text.parse::<f64>().unwrap_or(0.0)
-    }
-
-    // the token's exact integer, parsed digit by digit so no width can move
-    // it. Saturation is reported as a clamp, the wire's rule for a value
-    // outside what the reader can hold (§4).
-    fn to_integer(&self, is_signed: bool) -> (i64, bool) {
-        let token = &self.bytes[..self.length];
-        let mut i = 0usize;
-        let mut negative = false;
-        if i < token.len() && (token[i] == b'-' || token[i] == b'+') {
-            negative = token[i] == b'-';
-            i += 1;
-        }
-        let mut magnitude = 0u64;
-        let mut over = false;
-        while i < token.len() {
-            let digit = (token[i] - b'0') as u64;
-            if magnitude > (u64::MAX - digit) / 10 {
-                over = true;
-                break;
-            }
-            magnitude = magnitude * 10 + digit;
-            i += 1;
-        }
-        if !is_signed {
-            // -0 IS zero, and clamping it would report an event that did not
-            // happen; only a real negative magnitude is out of range here
-            if negative {
-                return (0, magnitude != 0);
-            }
-            if over {
-                return (u64::MAX as i64, true);
-            }
-            return (magnitude as i64, false);
-        }
-        if negative {
-            if over || magnitude > (1u64 << 63) {
-                return (i64::MIN, true);
-            }
-            if magnitude == (1u64 << 63) {
-                return (i64::MIN, false);
-            }
-            return (-(magnitude as i64), false);
-        }
-        if over || magnitude > i64::MAX as u64 {
-            return (i64::MAX, true);
-        }
-        (magnitude as i64, false)
-    }
-}
-
-fn table_json_skip_container(
-    input: &mut TableJsonIn,
-    close: u8,
-    depth: i32,
-    report: &mut TableReport,
-) -> bool {
-    if depth > TABLE_JSON_MAX_DEPTH {
-        input.bad = true;
-        return false;
-    }
-    input.pos += 1; // the opening bracket
-    loop {
-        let mut c = input.peek();
-        if c == close {
-            input.pos += 1;
-            return true;
-        }
-        if c == 0 {
-            input.bad = true;
-            return false;
-        }
-        if close == b'}' {
-            if table_json_scan_string(input, TableJsonSink::Discard, report).is_none() {
-                return false;
-            }
-            if input.peek() != b':' {
-                input.bad = true;
-                return false;
-            }
-            input.pos += 1;
-        }
-        if !table_json_skip_value(input, depth + 1, report) {
-            return false;
-        }
-        c = input.peek();
-        if c == b',' {
-            input.pos += 1; // a trailing comma is accepted
-            continue;
-        }
-        if c == close {
-            input.pos += 1;
-            return true;
-        }
-        input.bad = true;
-        return false;
-    }
-}
-
-fn table_json_skip_value(input: &mut TableJsonIn, depth: i32, report: &mut TableReport) -> bool {
-    match input.peek() {
-        b'{' => table_json_skip_container(input, b'}', depth, report),
-        b'[' => table_json_skip_container(input, b']', depth, report),
-        b'"' => table_json_scan_string(input, TableJsonSink::Discard, report).is_some(),
-        b't' => input.literal(b"true"),
-        b'f' => input.literal(b"false"),
-        b'n' => input.literal(b"null"),
-        0 => {
-            input.bad = true;
-            false
-        }
-        _ => {
-            // consumed, never converted: skipping needs no buffer, and this is
-            // the one walk a hostile text drives to the depth cap. It is the
-            // SAME production the value path scans, so an unknown key cannot
-            // smuggle past a number a named key would refuse.
-            if table_json_walk_number(input).is_none() {
-                input.bad = true;
-                return false;
-            }
-            true
-        }
-    }
-}
-
-// place one scalar at one storage address
-/// # Safety
-/// storage must point at the field's own slot.
-unsafe fn table_json_read_scalar(
-    input: &mut TableJsonIn,
-    storage: *mut u8,
-    f: &TableFieldInfo,
-    depth: i32,
-    report: &mut TableReport,
-) -> bool {
-    unsafe {
-        if let Some(arms_of) = f.arms {
-            // a union is an object with ONE key, the arm's name; {} is None,
-            // and two keys is a text this walk will not guess at
-            let arms = arms_of();
-            if input.peek() != b'{' {
-                input.bad = true;
-                return false;
-            }
-            input.pos += 1;
-            (arms.clear)(storage);
-            if input.peek() == b'}' {
-                input.pos += 1;
-                return true;
-            }
-            let mut key = TableJsonKey::new();
-            if !table_json_scan_key(input, &mut key, report) {
-                return false;
-            }
-            if input.peek() != b':' {
-                input.bad = true;
-                return false;
-            }
-            input.pos += 1;
-            let name_of = f.enum_name.unwrap();
-            let mut tag = 0i64;
-            for t in 1..=f.enum_max {
-                if name_of(t as u64) == key.as_str() {
-                    tag = t;
-                    break;
-                }
-            }
-            if tag == 0 {
-                report.unknown += 1;
-                if !table_json_skip_value(input, depth + 1, report) {
-                    return false;
-                }
-            } else {
-                let table = arms.arms[tag as usize].table.unwrap()();
-                let payload = (arms.select)(storage, tag as u64);
-                if !table_json_read_table(input, payload, table, depth + 1, report) {
-                    return false;
-                }
-            }
-            let mut c = input.peek();
-            if c == b',' {
-                input.pos += 1;
-                c = input.peek();
-            }
-            if c == b'}' {
-                input.pos += 1;
-                return true;
-            }
-            input.bad = true; // a second key: a one-of with two arms is not a value
-            return false;
-        }
-        if f.kind == 13 {
-            let table = f.table.unwrap()();
-            (table.reset)(storage);
-            return table_json_read_table(input, storage, table, depth + 1, report);
-        }
-        if table_json_is_enum(f) {
-            let mut name = TableJsonKey::new();
-            if !table_json_scan_key(input, &mut name, report) {
-                return false;
-            }
-            let name_of = f.enum_name.unwrap();
-            for v in 0..=f.enum_max {
-                if name_of(v as u64) == name.as_str() {
-                    table_json_set_raw(storage, f.elem_size, v as u64);
-                    return true;
-                }
-            }
-            // a name this build cannot name reads as None and counts as
-            // unknown, exactly as an unknown variant id does on the wire (§4)
-            table_json_set_raw(storage, f.elem_size, 0);
-            report.unknown += 1;
-            return true;
-        }
-        if table_json_is_flags(f) {
-            if input.peek() != b'[' {
-                input.bad = true;
-                return false;
-            }
-            input.pos += 1;
-            let flag_name = f.flag_name.unwrap();
-            let mut bits = 0u64;
-            loop {
-                let mut c = input.peek();
-                if c == b']' {
-                    input.pos += 1;
-                    break;
-                }
-                if c == 0 {
-                    input.bad = true;
-                    return false;
-                }
-                if c != b'"' {
-                    report.kind_mismatch += 1;
-                    if !table_json_skip_value(input, depth + 1, report) {
-                        return false;
-                    }
-                } else {
-                    let mut name = TableJsonKey::new();
-                    if !table_json_scan_key(input, &mut name, report) {
-                        return false;
-                    }
-                    let mut found = false;
-                    for bit in 0..=f.enum_max {
-                        if flag_name(bit as u32) == name.as_str() {
-                            bits |= 1u64 << bit;
-                            found = true;
-                            break;
-                        }
-                    }
-                    if !found {
-                        report.unknown += 1;
-                    }
-                }
-                c = input.peek();
-                if c == b',' {
-                    input.pos += 1;
-                    continue;
-                }
-                if c == b']' {
-                    input.pos += 1;
-                    break;
-                }
-                input.bad = true;
-                return false;
-            }
-            table_json_set_raw(storage, f.elem_size, bits);
-            return true;
-        }
-        if f.kind == 1 {
-            let c = input.peek();
-            if c == b't' {
-                if !input.literal(b"true") {
-                    return false;
-                }
-                table_json_set_raw(storage, f.elem_size, 1);
-                return true;
-            }
-            if !input.literal(b"false") {
-                return false;
-            }
-            table_json_set_raw(storage, f.elem_size, 0);
-            return true;
-        }
-        let token = match table_json_scan_number(input) {
-            Some(t) => t,
-            None => {
-                input.bad = true;
-                return false;
-            }
-        };
-        if f.kind == 10 || f.kind == 11 {
-            let single = f.kind == 10;
-            let mut value = token.to_double(single);
-            // A magnitude the field's format cannot hold is the WRONG SHAPE
-            // for the kind, and it never reaches storage: 1e400 is not a
-            // float64 and 1e300 is not a float32. §16.1's one invariant is
-            // that a text which reads clean writes back.
-            if !table_json_finite(value) {
-                report.kind_mismatch += 1;
-                return true;
-            }
-            if f.has_range {
-                if value < f.range_min {
-                    value = f.range_min;
-                    report.clamped += 1;
-                } else if value > f.range_max {
-                    value = f.range_max;
-                    report.clamped += 1;
-                }
-            }
-            if single {
-                let narrow = value as f32;
-                if !table_json_finite(narrow as f64) {
-                    report.kind_mismatch += 1;
-                    return true;
-                }
-                table_json_set_raw(storage, 4, narrow.to_bits() as u64);
-            } else {
-                table_json_set_raw(storage, 8, value.to_bits());
-            }
-            return true;
-        }
-        // JSON HAS ONE NUMBER TYPE. 2.0 IS the integer 2 and 1e3 IS 1000, and
-        // a library that round-trips numbers through a double emits them that
-        // way — this walker's own float writer emits 1e+21. So an integer
-        // field takes any number whose VALUE is integral, however it was
-        // spelled; only a genuinely fractional value is the wrong shape.
-        let is_signed = (2..=5).contains(&f.kind);
-        let mut saturated;
-        let mut value;
-        if token.integral {
-            let (v, s) = token.to_integer(is_signed);
-            value = v;
-            saturated = s;
-        } else {
-            let d = token.to_double(false);
-            if !table_json_finite(d) {
-                report.kind_mismatch += 1;
-                return true;
-            }
-            saturated = false;
-            if is_signed {
-                if d >= 9223372036854775808.0 {
-                    value = i64::MAX;
-                    saturated = true;
-                } else if d < -9223372036854775808.0 {
-                    value = i64::MIN;
-                    saturated = true;
-                } else if d != (d as i64) as f64 {
-                    report.kind_mismatch += 1;
-                    return true;
-                } else {
-                    value = d as i64;
-                }
-            } else if d < 0.0 {
-                // a negative for an unsigned field clamps to zero, as the
-                // exact digit path already does
-                if d != (d as i64) as f64 {
-                    report.kind_mismatch += 1;
-                    return true;
-                }
-                value = 0;
-                saturated = true;
-            } else if d >= 18446744073709551616.0 {
-                value = u64::MAX as i64;
-                saturated = true;
-            } else if d != (d as u64) as f64 {
-                report.kind_mismatch += 1;
-                return true;
-            } else {
-                value = (d as u64) as i64;
-            }
-        }
-        if saturated {
-            report.clamped += 1;
-        }
-        if f.has_range {
-            if (value as f64) < f.range_min {
-                value = f.range_min as i64;
-                report.clamped += 1;
-            } else if (value as f64) > f.range_max {
-                value = f.range_max as i64;
-                report.clamped += 1;
-            }
-        }
-        // the field's own storage width is the last bound: a value past it
-        // clamps rather than wrapping, which is what the wire does too
-        if f.elem_size < 8 {
-            if is_signed {
-                let high = (1i64 << (f.elem_size * 8 - 1)) - 1;
-                let low = -high - 1;
-                if value > high {
-                    value = high;
-                    report.clamped += 1;
-                } else if value < low {
-                    value = low;
-                    report.clamped += 1;
-                }
-            } else {
-                let high = (1u64 << (f.elem_size * 8)) - 1;
-                if value < 0 {
-                    value = 0;
-                    report.clamped += 1;
-                } else if value as u64 > high {
-                    value = high as i64;
-                    report.clamped += 1;
-                }
-            }
-        }
-        // at eight bytes the storage IS the parser's width, and an unsigned
-        // value past i64::MAX rides here as a negative i64 by design.
-        table_json_set_raw(storage, f.elem_size, value as u64);
-        true
-    }
-}
-
-/// # Safety
-/// base must point at an instance of the record the descriptor describes.
-unsafe fn table_json_read_field(
-    input: &mut TableJsonIn,
-    base: *mut u8,
-    f: &TableFieldInfo,
-    depth: i32,
-    report: &mut TableReport,
-) -> bool {
-    unsafe {
-        let storage = base.add(f.offset as usize);
-        if f.kind == 12 {
-            let sink = TableJsonSink::Storage(storage, f.array_bound as usize);
-            let length = match table_json_scan_string(input, sink, report) {
-                Some(n) => n,
-                None => return false,
-            };
-            table_json_set_count(base, f, length);
-            return true;
-        }
-        if table_json_is_bytes(f) {
-            // base64 decodes STRAIGHT INTO the field's storage, six bits at a
-            // time — no window, no temporary, so a bytes(N) of any declared
-            // extent reads the same way. A base64 body carries no escapes, so
-            // a backslash in one is simply not an alphabet character.
-            if input.peek() != b'"' {
-                input.bad = true;
-                return false;
-            }
-            input.pos += 1;
-            let mut pos = input.pos;
-            let mut pad = 0i32;
-            let mut symbols = 0i32;
-            let mut malformed = false;
-            loop {
-                if pos >= input.text.len() {
-                    input.bad = true;
-                    return false;
-                }
-                let c = input.text[pos];
-                pos += 1;
-                if c == b'"' {
-                    break;
-                }
-                if c == b'=' {
-                    pad += 1;
-                    continue;
-                }
-                if pad > 0 {
-                    malformed = true;
-                    continue;
-                }
-                if c >= 128 || TABLE_JSON_BASE64_DECODE[c as usize] < 0 {
-                    malformed = true;
-                    continue;
-                }
-                symbols += 1;
-            }
-            if malformed
-                || (pad > 0 && ((symbols + pad) % 4 != 0 || symbols % 4 < 2 || pad > 2))
-                || (pad == 0 && symbols % 4 == 1)
-            {
-                input.pos = pos;
-                report.kind_mismatch += 1;
-                return true;
-            }
-            ptr::write_bytes(storage, 0, f.array_bound as usize);
-            table_json_set_count(base, f, 0);
-            let mut placed = 0i32;
-            let mut accumulator = 0u32;
-            let mut held = 0i32;
-            let mut clamped = false;
-            while input.pos < pos {
-                let c = input.text[input.pos];
-                input.pos += 1;
-                if c == b'"' || c == b'=' {
-                    break;
-                }
-                let at = TABLE_JSON_BASE64_DECODE[c as usize];
-                accumulator = (accumulator << 6) | at as u32;
-                held += 6;
-                if held >= 8 {
-                    held -= 8;
-                    if placed < f.array_bound {
-                        ptr::write(
-                            storage.add(placed as usize),
-                            ((accumulator >> held) & 0xff) as u8,
-                        );
-                        placed += 1;
-                    } else {
-                        clamped = true;
-                    }
-                }
-            }
-            input.pos = pos;
-            if clamped {
-                report.clamped += 1;
-            }
-            table_json_set_count(base, f, placed);
-            return true;
-        }
-        if table_json_is_keyed(f) {
-            if input.peek() != b'{' {
-                input.bad = true;
-                return false;
-            }
-            input.pos += 1;
-            // every slot back to its declared defaults first, so a key the
-            // text omits keeps them and a repeated field key cannot leave an
-            // earlier occurrence's slots standing
-            for i in 0..f.array_bound as usize {
-                let slot = storage.add(i * f.elem_size as usize);
-                match f.table {
-                    Some(t) if f.kind == 13 => (t().reset)(slot),
-                    _ => ptr::write_bytes(slot, 0, f.elem_size as usize),
-                }
-            }
-            let shape = table_json_element_shape(f);
-            // A KEYED OBJECT'S KEYS ARE KEYS: a variant named twice is a
-            // duplicate key like any other, last-wins and counted (§16.2).
-            let mut seen = [0u64; 8];
-            loop {
-                let mut c = input.peek();
-                if c == b'}' {
-                    input.pos += 1;
-                    break;
-                }
-                if c == 0 {
-                    input.bad = true;
-                    return false;
-                }
-                let mut key = TableJsonKey::new();
-                if !table_json_scan_key(input, &mut key, report) {
-                    return false;
-                }
-                if input.peek() != b':' {
-                    input.bad = true;
-                    return false;
-                }
-                input.pos += 1;
-                let mut slot: i64 = -1;
-                for v in 0..f.array_bound as i64 {
-                    // nothing is stored for None, so "None" finds no slot and
-                    // is an unknown key like any other name this reader
-                    // cannot place
-                    if !table_json_keyed_slot_valid(f, v) {
-                        continue;
-                    }
-                    if f.key_name.unwrap()(table_json_keyed_slot_key(v)) == key.as_str() {
-                        slot = v;
-                        break;
-                    }
-                }
-                if slot >= 0 && slot < 512 {
-                    let bit = 1u64 << (slot & 63);
-                    if seen[(slot >> 6) as usize] & bit != 0 {
-                        report.duplicate += 1;
-                    }
-                    seen[(slot >> 6) as usize] |= bit;
-                }
-                if slot < 0 {
-                    report.unknown += 1;
-                    if !table_json_skip_value(input, depth + 1, report) {
-                        return false;
-                    }
-                } else if input.value_shape() != shape {
-                    report.kind_mismatch += 1;
-                    if !table_json_skip_value(input, depth + 1, report) {
-                        return false;
-                    }
-                } else if !table_json_read_scalar(
-                    input,
-                    storage.add(slot as usize * f.elem_size as usize),
-                    f,
-                    depth + 1,
-                    report,
-                ) {
-                    return false;
-                }
-                c = input.peek();
-                if c == b',' {
-                    input.pos += 1; // a trailing comma is accepted
-                    continue;
-                }
-                if c == b'}' {
-                    input.pos += 1;
-                    break;
-                }
-                input.bad = true;
-                return false;
-            }
-            return true;
-        }
-        if f.is_array {
-            if input.peek() != b'[' {
-                input.bad = true;
-                return false;
-            }
-            input.pos += 1;
-            // LAST WINS has to be true of a repeated ARRAY key too, and it is
-            // wire-visible: a fixed array writes every slot, so a second,
-            // shorter occurrence overlaying a prefix would leave the first
-            // occurrence's tail standing.
-            match f.table {
-                Some(t) if f.kind == 13 => {
-                    let info = t();
-                    for i in 0..f.array_bound as usize {
-                        (info.reset)(storage.add(i * f.elem_size as usize));
-                    }
-                }
-                _ => ptr::write_bytes(
-                    storage,
-                    0,
-                    f.array_bound as usize * f.elem_size as usize,
-                ),
-            }
-            table_json_set_count(base, f, 0);
-            let mut placed = 0i32;
-            let shape = table_json_element_shape(f);
-            loop {
-                let mut c = input.peek();
-                if c == b']' {
-                    input.pos += 1;
-                    break;
-                }
-                if c == 0 {
-                    input.bad = true;
-                    return false;
-                }
-                if placed >= f.array_bound {
-                    // more elements than the reader's bound: the bounded
-                    // prefix is kept and the excess counts, the wire's rule
-                    report.clamped += 1;
-                    if !table_json_skip_value(input, depth + 1, report) {
-                        return false;
-                    }
-                } else if input.value_shape() != shape {
-                    report.kind_mismatch += 1;
-                    if !table_json_skip_value(input, depth + 1, report) {
-                        return false;
-                    }
-                    placed += 1;
-                } else {
-                    if !table_json_read_scalar(
-                        input,
-                        storage.add(placed as usize * f.elem_size as usize),
-                        f,
-                        depth + 1,
-                        report,
-                    ) {
-                        return false;
-                    }
-                    placed += 1;
-                }
-                c = input.peek();
-                if c == b',' {
-                    input.pos += 1;
-                    continue;
-                }
-                if c == b']' {
-                    input.pos += 1;
-                    break;
-                }
-                input.bad = true;
-                return false;
-            }
-            // a fixed array's tail keeps the defaults the prefill left there,
-            // exactly as a short wire count does
-            table_json_set_count(base, f, placed);
-            return true;
-        }
-        table_json_read_scalar(input, storage, f, depth, report)
-    }
-}
-
-// ONE table object: keys are field keys, unknown ones are skipped and
-// counted, a repeated key is last-wins and counted. The instance is already
-// at its declared defaults when this is entered, so a key the text never
-// mentions keeps the default an absent field takes on the wire (§4).
-/// # Safety
-/// base must point at an instance of the record info describes.
-unsafe fn table_json_read_table(
-    input: &mut TableJsonIn,
-    base: *mut u8,
-    info: &TableTypeInfo,
-    depth: i32,
-    report: &mut TableReport,
-) -> bool {
-    unsafe {
-        if depth > TABLE_JSON_MAX_DEPTH {
-            input.bad = true;
-            return false;
-        }
-        if input.peek() != b'{' {
-            input.bad = true;
-            return false;
-        }
-        input.pos += 1;
-        // duplicate tracking, bounded and allocation-free: a table with more
-        // fields than this still reads, its repeats simply stop being counted
-        let mut seen = [0u64; 8];
-        loop {
-            let mut c = input.peek();
-            if c == b'}' {
-                input.pos += 1;
-                return true;
-            }
-            if c == 0 {
-                input.bad = true;
-                return false;
-            }
-            let mut key = TableJsonKey::new();
-            if !table_json_scan_key(input, &mut key, report) {
-                return false;
-            }
-            if input.peek() != b':' {
-                input.bad = true;
-                return false;
-            }
-            input.pos += 1;
-            let mut index: i32 = -1;
-            for (i, f) in info.fields.iter().enumerate() {
-                if f.json == key.as_str() {
-                    index = i as i32;
-                    break;
-                }
-            }
-            if key.as_str().starts_with('&') {
-                // THE AMPERSAND PREFIX IS RESERVED TO THE FORM (docs/SPEC-TABLES.md
-                // 16.7): never a field this build lacks, always a construct it
-                // cannot honor. MALFORMED and refused, never counted as unknown.
-                report.malformed = true;
-                input.bad = true;
-                return false;
-            }
-            if index < 0 {
-                report.unknown += 1;
-                if !table_json_skip_value(input, depth + 1, report) {
-                    return false;
-                }
-            } else {
-                let f = &info.fields[index as usize];
-                if index < 512 {
-                    let bit = 1u64 << (index & 63);
-                    if seen[(index >> 6) as usize] & bit != 0 {
-                        report.duplicate += 1;
-                    }
-                    seen[(index >> 6) as usize] |= bit;
-                }
-                // PRESENCE OF THE KEY IS THE PRESENCE (§16.2): reaching this
-                // line is the key being present, so an optional is set
-                // present whatever its value — with one exception the page
-                // names: a JSON null, which reads as ABSENT.
-                let got = input.value_shape();
-                if f.optional && got == b'z' {
-                    if !input.literal(b"null") {
-                        return false;
-                    }
-                    // absent, and back at its defaults: a repeated key whose
-                    // last occurrence is null must not leave an earlier value
-                    // standing
-                    match f.table {
-                        Some(t) => (t().reset)(base.add(f.offset as usize)),
-                        None => ptr::write_bytes(
-                            base.add(f.offset as usize),
-                            0,
-                            f.elem_size as usize,
-                        ),
-                    }
-                    table_json_set_raw(base.add(f.present_offset as usize), 1, 0);
-                } else {
-                    if got != table_json_shape(f) {
-                        // the wrong JSON type for the kind: skipped, never coerced
-                        report.kind_mismatch += 1;
-                        if !table_json_skip_value(input, depth + 1, report) {
-                            return false;
-                        }
-                    } else if !table_json_read_field(input, base, f, depth, report) {
-                        return false;
-                    }
-                    if f.optional {
-                        table_json_set_raw(base.add(f.present_offset as usize), 1, 1);
-                    }
-                }
-            }
-            c = input.peek();
-            if c == b',' {
-                input.pos += 1; // a trailing comma is accepted
-                continue;
-            }
-            if c == b'}' {
-                input.pos += 1;
-                return true;
-            }
-            input.bad = true;
-            return false;
-        }
-    }
-}
-
-// ---- the two entry points the per-table wrappers name ----
-
-/// # Safety
-/// value must point at an instance of the record info describes.
-pub unsafe fn table_json_read(
-    value: *mut u8,
-    info: &TableTypeInfo,
-    text: &[u8],
-    report: &mut TableReport,
-) -> bool {
-    unsafe {
-        let mut input = TableJsonIn {
-            text,
-            pos: 0,
-            bad: false,
-        };
-        (info.reset)(value);
-        let ok = table_json_read_table(&mut input, value, info, 0, report);
-        if ok {
-            input.space();
-            if input.pos != input.text.len() {
-                input.bad = true; // trailing rubbish is not one text
-            }
-        }
-        if input.bad || !ok {
-            report.malformed = true;
-            return false;
-        }
-        true
-    }
-}
-
-/// # Safety
-/// value must point at an instance of the record info describes.
-pub unsafe fn table_json_write(
-    value: *const u8,
-    info: &TableTypeInfo,
-    buffer: Option<&mut [u8]>,
-) -> i64 {
-    unsafe {
-        let mut out = TableJsonOut {
-            buffer,
-            offset: 0,
-            overflow: false,
-        };
-        if !table_json_write_value(&mut out, value, info, 0) {
-            return -1;
-        }
-        // THE CANONICAL TEXT ENDS WITH EXACTLY ONE NEWLINE (docs/SPEC-TABLES.md
-        // §16.1). Every writer emits it — this walk, the C++ walk, the C# walk
-        // and "schema unpack" — and every reader accepts a text with or without
-        // one, because the trailing whitespace a read already skips is what
-        // makes the two the same text. It is a byte of the FORM rather than a
-        // file convention: a text that is written to a file, pasted into a diff
-        // and handed back through a pipe has to be one text in all three
-        // places, and a buffer whose last byte is a closing brace is the one
-        // shape that is not.
-        out.put(b'\n');
-        if out.overflow {
-            return -1;
-        }
-        out.offset
-    }
-}
-
-// ---- json walk: end ----
-`
-
-// unused keeps the fmt import honest while the module grows.
-var _ = fmt.Sprintf
