+++ internal/codegen/rusttable/rusttable.go
@@ -1,35 +1,15 @@
-// Package rusttable emits the Rust TABLE surface (docs/SPEC-TABLES.md), mirroring
-// internal/codegen/cpptable — the reference — the way internal/codegen/cstable
-// does. One module per unit file that declares tables (<base>_table.rs), one
-// shared runtime module per unit (table_runtime.rs), and the two ACCELERATORS
-// beside them: the block form's <base>_block.rs and the cooked form's read
-// side.
+// Package rusttable emits the Rust table accelerators (docs/SPEC-TABLES.md):
+// the block form's <base>_block.rs and the cooked form's read side, with
+// their shared runtimes (block_runtime.rs and cook_runtime.rs) and the unit's
+// build_version.rs.
 //
-// The wire is neutral, evolution-tolerant TLV: field identity is the name
-// hash, unknown fields skip, absent fields default, changed kinds skip (never
-// misdecode), out-of-range values clamp, framing damage stops the decode with
-// a partial result — and every event lands in the TableReport. Plain byte
-// code with no serialize dependency, so a table module compiles into any
-// crate; the encode surface is a measure/save split, so a caller can measure
-// nested tables in parallel, prefix-sum offsets, and scatter-write disjoint
-// ranges from N workers. Generated codecs allocate nothing: the caller owns
-// every buffer.
-//
-// TWO DEVIATIONS FROM THE C++ REFERENCE, both forced by the language and both
-// named here rather than discovered in the source:
-//
-//   - THE REPORT IS NOT INSIDE THE READER. C++'s TableReader carries a
-//     TableReport pointer; a Rust reader carrying `&mut TableReport` could
-//     not hand a sub-reader out of its own buffer while that borrow stood.
-//     The report is therefore the codecs' second parameter, threaded down
-//     beside the reader. One report, one caller, same counters.
-//   - THE COOKED AND BLOCKED RECORDS ARE THEIR OWN <Name>Row STRUCTS, as
-//     they are in C#. C++'s cooked record IS the table's storage struct
-//     because both spell `string(N)` as `char[N + 1]`; Rust's storage column
-//     (SPEC §6.1) spells it `[u8; N]`, so the two families cannot be one
-//     without moving the whole Rust target's storage. The layout model
-//     (ir.RecordLayout) is the contract, and the Row structs meet it, asserted
-//     at COMPILE TIME with const asserts over core::mem::offset_of!.
+// THE COOKED AND BLOCKED RECORDS ARE THEIR OWN <Name>Row STRUCTS, as they are
+// in C#. C++'s cooked record IS the table's storage struct because both spell
+// `string(N)` as `char[N + 1]`; Rust's storage column (SPEC §6.1) spells it
+// `[u8; N]`, so the two families cannot be one without moving the whole Rust
+// target's storage. The layout model (ir.RecordLayout) is the contract, and the
+// Row structs meet it, asserted at COMPILE TIME with const asserts over
+// core::mem::offset_of!.
 package rusttable
 
 import (
@@ -43,41 +23,18 @@ import (
 	"github.com/mas-bandwidth/schema/v2/ir"
 )
 
-// table-wire kinds (docs/SPEC-TABLES.md §3), named locally over the one
-// target-independent definition in ir — the vocabulary is wire law.
+// BlockRuntimeModule, CookRuntimeModule and BuildVersionModule are the shared
+// modules a unit with tables grows.
 const (
-	tkBool  = ir.TableKindBool
-	tkI8    = ir.TableKindI8
-	tkI16   = ir.TableKindI16
-	tkI32   = ir.TableKindI32
-	tkI64   = ir.TableKindI64
-	tkU8    = ir.TableKindU8
-	tkU16   = ir.TableKindU16
-	tkU32   = ir.TableKindU32
-	tkU64   = ir.TableKindU64
-	tkF32   = ir.TableKindF32
-	tkF64   = ir.TableKindF64
-	tkTable = ir.TableKindTable
-	tkUnion = ir.TableKindUnion
-)
-
-// RuntimeModule and BlockRuntimeModule are the two SHARED modules a unit with
-// tables grows. They are module names, so they are claimed the way the crate
-// root is: a schema file lowering to one of them would be silently replaced.
-const (
-	RuntimeModule      = "table_runtime"
 	BlockRuntimeModule = "block_runtime"
 	CookRuntimeModule  = "cook_runtime"
-	// BuildVersionModule is the unit's BUILD VERSION and nothing else. It is
-	// its own module because §20 says what it is: one digest answering "which
-	// build?" and not "which form?", so it belongs to neither accelerator.
-	// Both of them compare against it, and either may be switched off.
 	BuildVersionModule = "build_version"
 )
 
 // Generate returns module filename -> file contents for the unit's table
-// surface. Empty when the unit declares no table: a table-free unit's Rust
-// output is byte-identical with this backend in the chain or out of it.
+// surface (block form, cooked form, and build version). Empty when the unit
+// declares no table: a table-free unit's Rust output is byte-identical with
+// this backend in the chain or out of it.
 func Generate(u *ir.Unit) (map[string][]byte, error) {
 	if len(u.Tables) == 0 {
 		return nil, nil
@@ -87,44 +44,16 @@ func Generate(u *ir.Unit) (map[string][]byte, error) {
 	}
 	out := map[string][]byte{}
 
-	variable := ir.VariableTables(u)
 	closure := ir.TableClosure(u)
 	blocks := ir.Blocks(u)
 
-	// the WIRE surface is refused by name for a unit that declares a
-	// variable-length table, exactly as the C# backend refuses it
-	// (docs/SPEC-TABLES.md §11): the arena, the builder, the region and the
-	// node-table codec are a named follow-on (§15). The two ACCELERATORS need
-	// no codec — a block and a cook are POINTED AT, not parsed — so both are
-	// emitted in full.
-	var refused []string
-	for name := range u.Tables {
-		if variable[name] {
-			refused = append(refused, name)
-		}
-	}
-	sort.Strings(refused)
-
-	banner := ""
-	if len(refused) > 0 {
-		banner = refusalBanner(refused)
-	}
-
-	if len(refused) == 0 {
-		out[RuntimeModule+".rs"] = runtimeModule(u, closure)
-	}
 	if anyCookable(u, closure) {
-		out[CookRuntimeModule+".rs"] = append([]byte(banner), cookRuntimeModule(u)...)
+		out[CookRuntimeModule+".rs"] = cookRuntimeModule(u)
 	}
-	out[BuildVersionModule+".rs"] = append([]byte(banner), buildVersionModule(u)...)
+	out[BuildVersionModule+".rs"] = buildVersionModule(u)
 
 	for _, f := range u.Files {
-		g := &gen{unit: u, file: f, variable: variable, closure: closure, blocks: blocks, banner: banner}
-		if len(refused) == 0 {
-			if body := g.tableModule(); body != nil {
-				out[strings.ToLower(f.Base)+"_table.rs"] = body
-			}
-		}
+		g := &gen{unit: u, file: f, closure: closure, blocks: blocks}
 		if body := g.recordsModule(); body != nil {
 			out[strings.ToLower(f.Base)+"_records.rs"] = body
 		}
@@ -137,7 +66,7 @@ func Generate(u *ir.Unit) (map[string][]byte, error) {
 	// lives in its own modules so a consumer that never opens a block pays
 	// only for a module it does not call (docs/SPEC-TABLES.md §19).
 	if blocks != nil {
-		blockOut, err := generateBlocks(u, blocks, banner)
+		blockOut, err := generateBlocks(u, blocks, "")
 		if err != nil {
 			return nil, err
 		}
@@ -157,58 +86,32 @@ func Modules(out map[string][]byte) []string {
 	return names
 }
 
-func refusalBanner(refused []string) string {
-	var b strings.Builder
-	b.WriteString("// THE RUST WIRE SURFACE OF THIS UNIT IS REFUSED, BY NAME (docs/SPEC-TABLES.md §11).\n")
-	b.WriteString("//\n")
-	fmt.Fprintf(&b, "// It declares variable-length tables (%s), and the Rust table\n", englishList(refused))
-	b.WriteString("// backend's VARIABLE CLASS — the arena, the builder, the region and the node-table\n")
-	b.WriteString("// codec — is a named follow-on (§15). No <base>_table.rs is emitted for this unit,\n")
-	b.WriteString("// so a consumer reaching for measure, save or load gets a missing name from its own\n")
-	b.WriteString("// compiler, beside this file, which says why.\n")
-	b.WriteString("//\n")
-	b.WriteString("// What IS emitted is the two ACCELERATORS, because neither needs a codec: a block\n")
-	b.WriteString("// (§19) and a cook (§7) are pointed at, not parsed. A build that loads this unit's\n")
-	b.WriteString("// cooked assets is served in full; one that wants the tolerant wire is not, and\n")
-	b.WriteString("// runs the tool or the C++ backend for it.\n\n")
-	return b.String()
-}
-
-func englishList(names []string) string {
-	switch len(names) {
-	case 0:
-		return ""
-	case 1:
-		return names[0]
+// runtimeSourceName is the file the runtime is attributed to: the unit's
+// first file that declares a table, so the banner names a real schema file
+// rather than the unit key.
+func runtimeSourceName(u *ir.Unit) string {
+	names := make([]string, 0, len(u.Files))
+	for _, f := range u.Files {
+		if len(f.Tables) > 0 {
+			names = append(names, f.Base)
+		}
 	}
-	return strings.Join(names[:len(names)-1], ", ") + " and " + names[len(names)-1]
+	sort.Strings(names)
+	if len(names) == 0 {
+		return u.Package
+	}
+	return names[0]
 }
 
 type gen struct {
-	unit     *ir.Unit
-	file     *ir.File
-	variable map[string]bool
-	closure  map[string]bool
-	blocks   *ir.BlockUnit
-	banner   string
-
-	// owner is the closure member whose codec is being emitted. It decides
-	// how an enum-keyed array is REACHED: a `table` declaration's storage is
-	// this backend's own TableKeyed, and a closure `type`'s comes from the
-	// packet emitter, where the same declaration is a plain [T; E.Max].
-	owner *ir.Struct
+	unit    *ir.Unit
+	file    *ir.File
+	closure map[string]bool
+	blocks  *ir.BlockUnit
+	banner  string
 
 	body   strings.Builder
 	indent string
-}
-
-// keyedSlots is the expression a codec walks an enum-keyed array's storage
-// with, by STORAGE INDEX.
-func (g *gen) keyedSlots(f *ir.Field) string {
-	if g.owner != nil && g.owner.IsTable {
-		return "value." + f.Name + ".slots"
-	}
-	return "value." + f.Name
 }
 
 func (g *gen) pf(format string, args ...any) {
@@ -300,24 +203,6 @@ func rustFieldType(t ir.FieldType) string {
 	return "u8"
 }
 
-// scalarKindWidth is the payload width a table-wire kind carries.
-func tableKindWidth(kind int) int {
-	switch kind {
-	case tkBool, tkI8, tkU8:
-		return 1
-	case tkI16, tkU16:
-		return 2
-	case tkI32, tkU32, tkF32:
-		return 4
-	case tkI64, tkU64, tkF64:
-		return 8
-	}
-	return 0
-}
-
-func putFn(width int) string { return fmt.Sprintf("put%d", width*8) }
-func getFn(width int) string { return fmt.Sprintf("get%d", width*8) }
-
 // ---- literals ----
 
 // formatFloat renders a float literal at the storage type's own precision, so
@@ -352,18 +237,6 @@ func intLit(v *big.Int, typ string) string {
 		return "i64::MIN"
 	}
 	return s
-}
-
-// keyedType is the Rust spelling of an enum-keyed array's storage: the extent
-// comes from the key enum's own MAX and is named nowhere else (§2.4).
-func keyedType(f *ir.Field) string {
-	return fmt.Sprintf("TableKeyed<%s, { %s::MAX.0 as usize }>", rustFieldType(f.Type), f.KeyEnum)
-}
-
-// keyedTypeExpr is the same type in EXPRESSION position, where Rust wants the
-// turbofish.
-func keyedTypeExpr(f *ir.Field) string {
-	return fmt.Sprintf("TableKeyed::<%s, { %s::MAX.0 as usize }>", rustFieldType(f.Type), f.KeyEnum)
 }
 
 // arrayLen renders a field's array extent as a Rust expression — parenthesised,
@@ -418,3 +291,79 @@ func unionOf(f *ir.Field) *ir.Union {
 	un, _ := f.Type.Ref.(*ir.Union)
 	return un
 }
+
+func orderTables(tables []*ir.Struct) []*ir.Struct {
+	n := len(tables)
+	byName := map[string]int{}
+	for i, st := range tables {
+		byName[st.Name] = i
+	}
+	adj := make([][]int, n)
+	indeg := make([]int, n)
+	for i, st := range tables {
+		for _, f := range st.Fields {
+			if f.Type.Kind != ir.TNamed {
+				continue
+			}
+			if ref, ok := f.Type.Ref.(*ir.Struct); ok && ref.IsTable {
+				if j, ok := byName[ref.Name]; ok && j != i {
+					adj[j] = append(adj[j], i)
+					indeg[i]++
+				}
+			}
+		}
+	}
+	order := make([]*ir.Struct, 0, n)
+	done := make([]bool, n)
+	for len(order) < n {
+		pick := -1
+		for i := range n {
+			if !done[i] && indeg[i] == 0 {
+				pick = i
+				break
+			}
+		}
+		if pick == -1 {
+			for i := range n {
+				if !done[i] {
+					pick = i
+					break
+				}
+			}
+		}
+		done[pick] = true
+		order = append(order, tables[pick])
+		for _, t := range adj[pick] {
+			indeg[t]--
+		}
+	}
+	return order
+}
+
+// tableFieldTypeName is the schema type name a descriptor and a comment
+// carry for one field.
+func tableFieldTypeName(f *ir.Field) string {
+	switch f.Type.Kind {
+	case ir.TBool:
+		return "bool"
+	case ir.TFloat32:
+		return "float32"
+	case ir.TFloat64:
+		return "float64"
+	case ir.TString:
+		return "string"
+	case ir.TBytes:
+		return "bytes"
+	case ir.TBits:
+		return fmt.Sprintf("bits(%d)", f.Type.Width)
+	case ir.TInt:
+		if f.Type.Signed {
+			return fmt.Sprintf("int%d", f.Type.Width)
+		}
+		return fmt.Sprintf("uint%d", f.Type.Width)
+	case ir.TNamed:
+		return f.Type.Name
+	}
+	return "?"
+}
+
