+++ internal/codegen/rusttable/rusttable_test.go
@@ -100,13 +100,10 @@ type Point
 	}
 }
 
-// TestSharedRuntimeIsUnitIndependent: the text form is ONE walk over the
-// descriptors (docs/SPEC-TABLES.md §16.1), so the shared runtime's bytes must
-// not vary with what a unit declares. The Makefile's tables-rust-walk holds
-// the whole corpus to this; here it is held between two units that disagree
-// about everything a schema can disagree about.
-func TestSharedRuntimeIsUnitIndependent(t *testing.T) {
-	one := generate(t, valueOnly)[RuntimeModule+".rs"]
+// TestBlockRuntimeIsUnitIndependent: the block runtime's bytes must not vary
+// with what a unit declares.
+func TestBlockRuntimeIsUnitIndependent(t *testing.T) {
+	one := generate(t, valueOnly)[BlockRuntimeModule+".rs"]
 	two := generate(t, `package other
 
 flags Perks { Shielded, Cloaked }
@@ -116,112 +113,60 @@ type Buff
     multiplier float32 = 1.0
 }
 
-union Effect
-{
-    buff Buff
-}
-
 table Wide
 {
     blob   bytes(16)
     perks  Perks
-    effect Effect
+    buff   Buff
     tally  [4]uint16
 }
-`)[RuntimeModule+".rs"]
+`)[BlockRuntimeModule+".rs"]
 	if len(one) == 0 || len(two) == 0 {
-		t.Fatal("a unit with tables emitted no shared runtime")
+		t.Fatal("a unit with tables emitted no block runtime")
 	}
-	// the banner names the schema file and the package; the runtime below it
-	// is the comparison
 	body := func(data []byte) string {
 		text := string(data)
-		if i := strings.Index(text, "// ---- the read report"); i >= 0 {
+		if i := strings.Index(text, "// The block form's shared runtime"); i >= 0 {
 			return text[i:]
 		}
 		return text
 	}
 	if body(one) != body(two) {
-		t.Error("the shared table runtime is not the same bytes in two units — the text form would be a codec per unit rather than one walk (§16.1)")
+		t.Error("the block runtime is not the same bytes in two units")
 	}
 }
 
-// TestPointeredUnitRefusesTheWireSurface: the variable class's refusal is a
-// refusal of the WIRE surface and of nothing else (§11). The two ACCELERATORS
-// need no codec, so both are emitted, and every file that IS emitted opens
-// with a banner naming the refused tables and the follow-on.
-func TestPointeredUnitRefusesTheWireSurface(t *testing.T) {
-	out := generate(t, pointered)
-	for name, data := range out {
-		if strings.HasSuffix(name, "_table.rs") || name == RuntimeModule+".rs" {
-			t.Errorf("a pointered unit emitted the wire surface %s", name)
-		}
-		if !strings.Contains(string(data), "REFUSED, BY NAME") {
-			t.Errorf("%s carries no refusal banner", name)
-		}
-		if !strings.Contains(string(data), "Node") {
-			t.Errorf("%s does not name the refused table", name)
-		}
-	}
-	if _, ok := out["probe_cook.rs"]; !ok {
-		t.Error("a pointered unit emitted no cook reader — a root is any table (§7)")
-	}
-}
-
-// TestValueOnlyUnitCarriesTheWholeSurface: what a fixed-class unit gets, by
-// name — the storage, the three codecs, the descriptors, the text form, and
-// the two accelerators.
-func TestValueOnlyUnitCarriesTheWholeSurface(t *testing.T) {
+// TestAcceleratorsEmitted: what a table unit gets — the two accelerators
+// (block and cook), their shared runtimes, records, and the build version.
+// No wire surface (*_table.rs or table_runtime.rs) is emitted.
+func TestAcceleratorsEmitted(t *testing.T) {
 	out := generate(t, valueOnly)
-	table := string(out["probe_table.rs"])
-	if table == "" {
-		t.Fatal("no probe_table.rs")
+	for _, unwanted := range []string{"probe_table.rs", "table_runtime.rs"} {
+		if _, ok := out[unwanted]; ok {
+			t.Errorf("emitted dead wire surface %s", unwanted)
+		}
 	}
 	for _, want := range []string{
-		"pub struct Config", "pub fn config_measure", "pub fn config_save",
-		"pub fn config_load", "pub fn config_reset", "pub fn config_table_type",
-		"pub fn config_from_json", "pub fn config_to_json", "pub fn config_to_json_measure",
-		"impl TableEnum for Grade", "TableKeyed<i32, { Grade::MAX.0 as usize }>",
-		"pub extra_present: bool",
-	} {
-		if !strings.Contains(table, want) {
-			t.Errorf("the fixed-class surface is missing %q", want)
-		}
-	}
-	// the generated codecs allocate nothing: the caller owns every buffer
-	for _, banned := range []string{"Vec<", "String::", "Box<", "to_string()", "format!"} {
-		if strings.Contains(table, banned) {
-			t.Errorf("the generated table module contains %q — generated codecs must not allocate", banned)
-		}
-	}
-	// and it stands alone: no serialize dependency reaches the table surface
-	for _, banned := range []string{"serialize::", "use serialize"} {
-		if strings.Contains(table, banned) {
-			t.Errorf("the generated table module contains %q — it must stand alone", banned)
-		}
-	}
-	for _, want := range []string{"probe_cook.rs", "probe_block.rs", "probe_records.rs",
+		"probe_cook.rs", "probe_block.rs", "probe_records.rs",
 		"block_runtime.rs", BuildVersionModule + ".rs",
-		CookRuntimeModule + ".rs", RuntimeModule + ".rs"} {
+		CookRuntimeModule + ".rs",
+	} {
 		if _, ok := out[want]; !ok {
-			t.Errorf("a fixed-class unit emitted no %s", want)
+			t.Errorf("a table unit emitted no %s", want)
 		}
 	}
 }
 
 // TestSharedRuntimesAreFileOrderIndependent is #347/#351's rule, held for
-// Rust: the unit's THREE shared runtimes are named by the PACKAGE, so a corpus
-// file that sorts earlier cannot relocate one of them. Before the cook runtime
-// moved out of "the first file that declares a table", adding a file named
-// earlier in the alphabet moved two thousand lines between modules — correct
-// output, and a diff nobody can read.
+// Rust: the unit's shared runtimes are named by the PACKAGE, so a corpus
+// file that sorts earlier cannot relocate one of them.
 func TestSharedRuntimesAreFileOrderIndependent(t *testing.T) {
 	one := generate(t, valueOnly)
 	two, err := Generate(unitWithExtraFile(t, valueOnly))
 	if err != nil {
 		t.Fatalf("generate: %v", err)
 	}
-	for _, name := range []string{RuntimeModule + ".rs", CookRuntimeModule + ".rs", BlockRuntimeModule + ".rs"} {
+	for _, name := range []string{CookRuntimeModule + ".rs", BlockRuntimeModule + ".rs"} {
 		a, ok := one[name]
 		if !ok {
 			t.Fatalf("no %s", name)
@@ -235,7 +180,7 @@ func TestSharedRuntimesAreFileOrderIndependent(t *testing.T) {
 		}
 	}
 	// and the per-file modules of the ORIGINAL file are untouched too
-	for _, name := range []string{"probe_table.rs", "probe_cook.rs", "probe_block.rs"} {
+	for _, name := range []string{"probe_cook.rs", "probe_block.rs", "probe_records.rs"} {
 		if string(one[name]) != string(two[name]) {
 			t.Errorf("%s moved when an unrelated file was added ahead of it", name)
 		}
@@ -276,22 +221,6 @@ func unitWithExtraFile(t *testing.T, src string) *ir.Unit {
 		t.Fatalf("check: %v", cerrs[0])
 	}
 	return u
-}
-
-// TestRelocatableStorageIsAsserted: every closure member is #[repr(C)] and
-// Copy, so a table value may be copied, mmapped or handed across a process
-// boundary as bytes (§9). The assert is a compile error the day one is not.
-func TestRelocatableStorageIsAsserted(t *testing.T) {
-	table := string(generate(t, valueOnly)["probe_table.rs"])
-	for _, want := range []string{
-		"const fn table_relocatable<T: Copy>()",
-		"const _: () = table_relocatable::<Config>();",
-		"const _: () = table_relocatable::<Inner>();",
-	} {
-		if !strings.Contains(table, want) {
-			t.Errorf("the relocatability assert is missing %q", want)
-		}
-	}
 }
 
 // TestCookLayoutIsAssertedAtCompileTime: the cooked record's C ABI layout is
