+++ test/conformance/rust/src/main.rs
@@ -16,50 +16,11 @@
 // a separate process to own, and one exec per surface is what keeps the leg
 // inside the two-minute rule (#320).
 
-use std::alloc::{GlobalAlloc, Layout, System, alloc_zeroed, dealloc};
+use std::alloc::{Layout, alloc_zeroed, dealloc};
 use std::fs;
-use std::hint::black_box;
 use std::io::Write;
 use std::path::Path;
 use std::process::exit;
-use std::sync::atomic::{AtomicI64, AtomicU64, Ordering};
-use std::time::{Duration, Instant};
-
-// ---------------------------------------------------------------------------
-// the allocation counter
-// ---------------------------------------------------------------------------
-//
-// The SOAK's instrument (docs/SPEC-TABLES.md: "every read path allocates
-// nothing"). Counting bytes at the global allocator is the honest measurement
-// — RSS answers a different question, and an allocator that grew by one byte
-// per iteration would take a very long time to show up in it. The counters are
-// relaxed atomics: they are read once per interval, never on a decision.
-struct Counting;
-
-static LIVE_BYTES: AtomicI64 = AtomicI64::new(0);
-static TOTAL_ALLOCATIONS: AtomicU64 = AtomicU64::new(0);
-
-unsafe impl GlobalAlloc for Counting {
-    unsafe fn alloc(&self, layout: Layout) -> *mut u8 {
-        LIVE_BYTES.fetch_add(layout.size() as i64, Ordering::Relaxed);
-        TOTAL_ALLOCATIONS.fetch_add(1, Ordering::Relaxed);
-        unsafe { System.alloc(layout) }
-    }
-
-    unsafe fn dealloc(&self, pointer: *mut u8, layout: Layout) {
-        LIVE_BYTES.fetch_sub(layout.size() as i64, Ordering::Relaxed);
-        unsafe { System.dealloc(pointer, layout) }
-    }
-
-    unsafe fn realloc(&self, pointer: *mut u8, layout: Layout, new_size: usize) -> *mut u8 {
-        LIVE_BYTES.fetch_add(new_size as i64 - layout.size() as i64, Ordering::Relaxed);
-        TOTAL_ALLOCATIONS.fetch_add(1, Ordering::Relaxed);
-        unsafe { System.realloc(pointer, layout, new_size) }
-    }
-}
-
-#[global_allocator]
-static ALLOCATOR: Counting = Counting;
 
 // ---------------------------------------------------------------------------
 // the manifest, read exactly as testdata/conformance/tables/FORMAT.md states it
@@ -121,225 +82,7 @@ fn spill(dir: &str, name: &str, data: &[u8]) {
     }
 }
 
-// ---------------------------------------------------------------------------
-// the codec table: one row per (unit, root) the corpus names
-// ---------------------------------------------------------------------------
-//
-// Each unit declares its OWN TableReport — the generated surface is one crate
-// per unit — so the driver carries one report shape of its own and each row
-// copies into it. Five counters is the whole of §4's report.
 
-#[derive(Default, Clone, Copy)]
-struct Report {
-    unknown: i32,
-    kind_mismatch: i32,
-    clamped: i32,
-    duplicate: i32,
-    malformed: bool,
-}
-
-struct Codec {
-    unit: &'static str,
-    root: &'static str,
-    // Load the wire bytes, then measure and save: the bytes a round trip
-    // produces, or None when the load refused.
-    wire: fn(&[u8], &mut Report) -> Option<Vec<u8>>,
-    // FromJson the text, then measure and save.
-    json_read: fn(&[u8], &mut Report) -> Option<Vec<u8>>,
-    // Load the wire bytes, then ToJson.
-    json_write: fn(&[u8], &mut Report) -> Option<Vec<u8>>,
-    // EVERY GENERATED PATH, over buffers the caller already owns, returning
-    // the number of allocations the GENERATED code made. The value is boxed
-    // and the buffer sized BEFORE the counter is snapshotted, so what the
-    // count measures is the codec and not this driver: "every read path
-    // allocates nothing" is a claim about the generated code, and an
-    // instrument that counted the driver's own Vec could never read zero.
-    audit: fn(&[u8], &[u8], &mut Vec<u8>) -> u64,
-}
-
-// One row, spelled once. The crate, the storage type and the six generated
-// entry points are all passed by path because Rust has no identifier
-// concatenation in macro_rules — which is the honest spelling anyway: every
-// name this driver calls is visible at the call site.
-macro_rules! codec {
-    ($unit:literal, $root:literal, $krate:ident, $ty:ty,
-     $measure:path, $save:path, $load:path,
-     $from_json:path, $to_json_measure:path, $to_json:path) => {
-        Codec {
-            unit: $unit,
-            root: $root,
-            wire: |bytes, out| {
-                // boxed: a table's storage is bounded but not small — the wide
-                // corpus's root is a quarter of a megabyte, and a driver that
-                // put one on the stack would be testing the stack
-                let mut value: Box<$ty> = Box::default();
-                let mut report = $krate::TableReport::default();
-                let ok = $load(&mut value, bytes, &mut report);
-                copy_report!(report, out);
-                if !ok {
-                    return None;
-                }
-                write_measured(|| $measure(&value), |buffer| $save(&value, buffer))
-            },
-            json_read: |text, out| {
-                let mut value: Box<$ty> = Box::default();
-                let mut report = $krate::TableReport::default();
-                let ok = $from_json(&mut value, text, &mut report);
-                copy_report!(report, out);
-                if !ok {
-                    return None;
-                }
-                write_measured(|| $measure(&value), |buffer| $save(&value, buffer))
-            },
-            json_write: |bytes, out| {
-                let mut value: Box<$ty> = Box::default();
-                let mut report = $krate::TableReport::default();
-                let ok = $load(&mut value, bytes, &mut report);
-                copy_report!(report, out);
-                if !ok {
-                    return None;
-                }
-                write_measured(
-                    || $to_json_measure(&value),
-                    |buffer| $to_json(&value, buffer),
-                )
-            },
-            audit: |wire, text, buffer| {
-                let mut value: Box<$ty> = Box::default();
-                let mut report = $krate::TableReport::default();
-                // the warm pass: the descriptors are built, the buffer is
-                // sized, and every one-time cost is paid before the count
-                // starts
-                $load(&mut value, wire, &mut report);
-                let wire_bytes = $measure(&value).max(0) as usize;
-                let text_bytes = $to_json_measure(&value).max(0) as usize;
-                buffer.resize(wire_bytes.max(text_bytes), 0);
-                let before = TOTAL_ALLOCATIONS.load(Ordering::Relaxed);
-                // THE MEASURED REGION: every generated read and write path,
-                // and not one allocation of this driver's own
-                let mut r = $krate::TableReport::default();
-                $load(&mut value, wire, &mut r);
-                black_box($measure(&value));
-                black_box($save(&value, buffer));
-                $from_json(&mut value, text, &mut r);
-                black_box($to_json_measure(&value));
-                black_box($to_json(&value, buffer));
-                if std::env::var_os("SOAK_SABOTAGE").is_some() {
-                    // THE NEGATIVE CONTROL: one allocation per iteration,
-                    // inside the measured region, so the gate has something to
-                    // find. A gate that has never fired proves nothing.
-                    black_box(Box::new(0u8));
-                }
-                TOTAL_ALLOCATIONS.load(Ordering::Relaxed) - before
-            },
-        }
-    };
-}
-
-macro_rules! copy_report {
-    ($from:expr, $to:expr) => {
-        $to.unknown = $from.unknown;
-        $to.kind_mismatch = $from.kind_mismatch;
-        $to.clamped = $from.clamped;
-        $to.duplicate = $from.duplicate;
-        $to.malformed = $from.malformed;
-    };
-}
-
-// write_measured is the measure/write symmetry, held: the buffer is exactly
-// the measured size and the write must fill it exactly.
-fn write_measured(measure: impl Fn() -> i64, write: impl Fn(&mut [u8]) -> i64) -> Option<Vec<u8>> {
-    let size = measure();
-    if size < 0 {
-        return None;
-    }
-    let mut buffer = vec![0u8; size as usize];
-    if write(&mut buffer) != size {
-        return None;
-    }
-    Some(buffer)
-}
-
-fn codecs() -> Vec<Codec> {
-    vec![
-        codec!(
-            "tabledemo", "RootConfig", tabledemo, tabledemo::RootConfig,
-            tabledemo::root_config_measure, tabledemo::root_config_save, tabledemo::root_config_load,
-            tabledemo::root_config_from_json, tabledemo::root_config_to_json_measure, tabledemo::root_config_to_json
-        ),
-        codec!(
-            "tabledemo", "ProfileConfig", tabledemo, tabledemo::ProfileConfig,
-            tabledemo::profile_config_measure, tabledemo::profile_config_save, tabledemo::profile_config_load,
-            tabledemo::profile_config_from_json, tabledemo::profile_config_to_json_measure, tabledemo::profile_config_to_json
-        ),
-        codec!(
-            "tabledemo", "LoadoutConfig", tabledemo, tabledemo::LoadoutConfig,
-            tabledemo::loadout_config_measure, tabledemo::loadout_config_save, tabledemo::loadout_config_load,
-            tabledemo::loadout_config_from_json, tabledemo::loadout_config_to_json_measure, tabledemo::loadout_config_to_json
-        ),
-        codec!(
-            "tabledemo", "WideBlob", tabledemo, tabledemo::WideBlob,
-            tabledemo::wide_blob_measure, tabledemo::wide_blob_save, tabledemo::wide_blob_load,
-            tabledemo::wide_blob_from_json, tabledemo::wide_blob_to_json_measure, tabledemo::wide_blob_to_json
-        ),
-        codec!(
-            "tabledemo", "ArchiveConfig", tabledemo, tabledemo::ArchiveConfig,
-            tabledemo::archive_config_measure, tabledemo::archive_config_save, tabledemo::archive_config_load,
-            tabledemo::archive_config_from_json, tabledemo::archive_config_to_json_measure, tabledemo::archive_config_to_json
-        ),
-        codec!(
-            "tabledemo", "KeyedConfig", tabledemo, tabledemo::KeyedConfig,
-            tabledemo::keyed_config_measure, tabledemo::keyed_config_save, tabledemo::keyed_config_load,
-            tabledemo::keyed_config_from_json, tabledemo::keyed_config_to_json_measure, tabledemo::keyed_config_to_json
-        ),
-        codec!(
-            "tabledemo", "PackConfig", tabledemo, tabledemo::PackConfig,
-            tabledemo::pack_config_measure, tabledemo::pack_config_save, tabledemo::pack_config_load,
-            tabledemo::pack_config_from_json, tabledemo::pack_config_to_json_measure, tabledemo::pack_config_to_json
-        ),
-        codec!(
-            "tblv1", "Cfg", tblv1, tblv1::Cfg,
-            tblv1::cfg_measure, tblv1::cfg_save, tblv1::cfg_load,
-            tblv1::cfg_from_json, tblv1::cfg_to_json_measure, tblv1::cfg_to_json
-        ),
-        codec!(
-            "tblv2", "Cfg", tblv2, tblv2::Cfg,
-            tblv2::cfg_measure, tblv2::cfg_save, tblv2::cfg_load,
-            tblv2::cfg_from_json, tblv2::cfg_to_json_measure, tblv2::cfg_to_json
-        ),
-        codec!(
-            "tblp1", "Chain", tblp1, tblp1::Chain,
-            tblp1::chain_measure, tblp1::chain_save, tblp1::chain_load,
-            tblp1::chain_from_json, tblp1::chain_to_json_measure, tblp1::chain_to_json
-        ),
-        codec!(
-            "tblp3", "Chain", tblp3, tblp3::Chain,
-            tblp3::chain_measure, tblp3::chain_save, tblp3::chain_load,
-            tblp3::chain_from_json, tblp3::chain_to_json_measure, tblp3::chain_to_json
-        ),
-    ]
-}
-
-/// The row for a (unit, root), or `None` where this backend has none: Rust
-/// refuses a pointered unit's wire by name (§11), so the cases over one are a
-/// missing FEATURE and the driver says so per case.
-fn codec_for<'a>(rows: &'a [Codec], unit: &str, root: &str) -> Option<&'a Codec> {
-    rows.iter().find(|c| c.unit == unit && c.root == root)
-}
-
-/// `spill_absent` says this backend cannot answer THIS CASE — a feature it
-/// lacks, not a test it failed. The harness counts it and the matrix prints it
-/// beside what the leg did answer (test/conformance/README.md).
-fn spill_absent(dir: &str, name: &str) {
-    spill(dir, &format!("{name}.absent"), &[]);
-}
-
-/// `no_text` marks an instance the corpus carries on the WIRE only — past the
-/// text form's depth cap by the form's own rule (docs/SPEC-TABLES.md §16.7) —
-/// so no leg is asked for its text.
-fn no_text(f: &[String]) -> bool {
-    f.len() > 5 && f[5] == "no-text"
-}
 
 // ---------------------------------------------------------------------------
 // aligned storage
@@ -577,82 +320,6 @@ fn open_block(name: &str, data: &[u8], extent: i64) -> bool {
 // the surfaces
 // ---------------------------------------------------------------------------
 
-fn surface_wire(manifest: &Manifest, out: &str) {
-    let rows = codecs();
-    for f in manifest.of_kind("instance") {
-        let Some(codec) = codec_for(&rows, &f[2], &f[3]) else {
-            spill_absent(out, &f[1]);
-            continue;
-        };
-        let wire = slurp(&f[4]);
-        let mut report = Report::default();
-        match (codec.wire)(&wire, &mut report) {
-            Some(bytes) => spill(out, &f[1], &bytes),
-            None => fail(&format!("{} did not round-trip", f[1])),
-        }
-    }
-}
-
-fn surface_report(manifest: &Manifest, out: &str) {
-    let rows = codecs();
-    for f in manifest.of_kind("report") {
-        let Some(codec) = codec_for(&rows, &f[2], &f[3]) else {
-            spill_absent(out, &f[1]);
-            continue;
-        };
-        let wire = slurp(&f[4]);
-        let mut report = Report::default();
-        let ok = (codec.wire)(&wire, &mut report).is_some();
-        let text = format!(
-            "{},{},{},{},{}\n",
-            report.unknown,
-            report.kind_mismatch,
-            report.clamped,
-            report.duplicate,
-            if report.malformed || !ok { "true" } else { "false" }
-        );
-        spill(out, &f[1], text.as_bytes());
-    }
-}
-
-fn surface_json_read(manifest: &Manifest, out: &str) {
-    let rows = codecs();
-    for f in manifest.of_kind("instance") {
-        if no_text(f) {
-            continue;
-        }
-        let Some(codec) = codec_for(&rows, &f[2], &f[3]) else {
-            spill_absent(out, &f[1]);
-            continue;
-        };
-        let path = format!("testdata/conformance/tables/json/{}.json", f[1]);
-        let text = slurp(&path);
-        let mut report = Report::default();
-        match (codec.json_read)(&text, &mut report) {
-            Some(bytes) => spill(out, &f[1], &bytes),
-            None => fail(&format!("{} did not read from its text", f[1])),
-        }
-    }
-}
-
-fn surface_json_write(manifest: &Manifest, out: &str) {
-    let rows = codecs();
-    for f in manifest.of_kind("instance") {
-        if no_text(f) {
-            continue;
-        }
-        let Some(codec) = codec_for(&rows, &f[2], &f[3]) else {
-            spill_absent(out, &format!("{}.json", f[1]));
-            continue;
-        };
-        let wire = slurp(&f[4]);
-        let mut report = Report::default();
-        match (codec.json_write)(&wire, &mut report) {
-            Some(text) => spill(out, &format!("{}.json", f[1]), &text),
-            None => fail(&format!("{} did not write its text", f[1])),
-        }
-    }
-}
 
 fn surface_block_dump(manifest: &Manifest, out: &str) {
     for f in manifest.of_kind("block") {
@@ -662,31 +329,6 @@ fn surface_block_dump(manifest: &Manifest, out: &str) {
     }
 }
 
-// json-hostile: one tree per rule the text form states (§16.2, §16.3, §17.5).
-// The answer is the REPORT the read produces, or `refused` — the same
-// two-valued verdict the engine's own gate holds, over the same data.
-fn surface_json_hostile(manifest: &Manifest, out: &str) {
-    let rows = codecs();
-    for f in manifest.of_kind("json-hostile") {
-        let Some(codec) = codec_for(&rows, &f[2], &f[3]) else {
-            spill_absent(out, &f[1]);
-            continue;
-        };
-        // the tree is what `schema pack` reads, so the text is <tree>/<root>.json (§17)
-        let text = slurp(&format!("{}/{}.json", f[4], f[3]));
-        let mut report = Report::default();
-        let ok = (codec.json_read)(&text, &mut report).is_some();
-        let verdict = if !ok || report.malformed {
-            "refused\n".to_string()
-        } else {
-            format!(
-                "{},{},{},{},false\n",
-                report.unknown, report.kind_mismatch, report.clamped, report.duplicate
-            )
-        };
-        spill(out, &f[1], verdict.as_bytes());
-    }
-}
 
 fn surface_block(manifest: &Manifest, out: &str) {
     for f in manifest.of_kind("block") {
@@ -1121,200 +763,6 @@ fn surface_cook(manifest: &Manifest, out: &str) {
 
 // ---------------------------------------------------------------------------
 
-// ---------------------------------------------------------------------------
-// the soak
-// ---------------------------------------------------------------------------
-//
-// Every instance of the corpus, wire-loaded and re-saved and text-read and
-// text-written, in a loop, for as long as the caller asks — with the bytes
-// compared every iteration, so a run that drifted would stop rather than merely
-// get slower, and with LIVE ALLOCATED BYTES printed each interval, which is the
-// property the read path owes: it allocates nothing beyond the value and the
-// buffers the caller passes.
-//
-// The caller's buffers are hoisted out of the loop on purpose. What the soak is
-// watching for is an allocation the GENERATED code makes, and a Vec allocated
-// per iteration by the driver would hide one under its own noise.
-fn surface_soak(manifest: &Manifest, seconds: u64) {
-    let rows = codecs();
-    struct Case<'a> {
-        name: &'a str,
-        codec: &'a Codec,
-        wire: Vec<u8>,
-        text: Vec<u8>,
-    }
-    let mut cases = Vec::new();
-    for f in manifest.of_kind("instance") {
-        // an instance this backend has no codec for, or no text for, is not
-        // one it can walk: the pointered class is ABSENT here (§11), and the
-        // corpus carries the deep chain on the wire alone (§16.7)
-        if no_text(f) {
-            continue;
-        }
-        let Some(codec) = codec_for(&rows, &f[2], &f[3]) else {
-            continue;
-        };
-        cases.push(Case {
-            name: &f[1],
-            codec,
-            wire: slurp(&f[4]),
-            text: slurp(&format!("testdata/conformance/tables/json/{}.json", f[1])),
-        });
-    }
-    if cases.is_empty() {
-        fail("the manifest names no instance to soak");
-    }
-
-    // one warm pass, so every lazily-built descriptor and every buffer this
-    // driver owns exists before the baseline is taken
-    let mut audit_buffer: Vec<u8> = Vec::new();
-    for case in cases.iter() {
-        let mut report = Report::default();
-        (case.codec.wire)(&case.wire, &mut report);
-        (case.codec.json_read)(&case.text, &mut report);
-        (case.codec.json_write)(&case.wire, &mut report);
-        (case.codec.audit)(&case.wire, &case.text, &mut audit_buffer);
-    }
-
-    // THE ALLOCATION GATE, and it is a different instrument from the drift
-    // one below. Live bytes answer "does this LEAK"; a path that allocated
-    // and freed the same bytes every iteration answers +0 there forever. The
-    // COUNT is what "every read path allocates nothing" actually claims, so
-    // it is gated separately and the first failing case names itself.
-    for case in cases.iter() {
-        let made = (case.codec.audit)(&case.wire, &case.text, &mut audit_buffer);
-        if made != 0 {
-            fail(&format!(
-                "{}: the generated code made {made} allocation(s) on the read and write paths — \
-                 docs/USAGE.md's claim is that it makes none beyond the value and the buffers \
-                 the caller passed",
-                case.name
-            ));
-        }
-    }
-
-    // the baseline is taken AFTER the first print, because stdout's own line
-    // buffer is allocated on first use and is a one-time runtime cost rather
-    // than anything the generated code did
-    println!(
-        "soak: {} instances, {seconds}s, live allocated bytes are the instrument",
-        cases.len()
-    );
-    let baseline = LIVE_BYTES.load(Ordering::Relaxed);
-    let allocations_baseline = TOTAL_ALLOCATIONS.load(Ordering::Relaxed);
-    let started = Instant::now();
-    let deadline = started + Duration::from_secs(seconds);
-    let mut next_report = started + Duration::from_secs(60.min(seconds));
-    let mut iterations: u64 = 0;
-
-    while Instant::now() < deadline {
-        for case in cases.iter() {
-            let mut report = Report::default();
-            match (case.codec.wire)(&case.wire, &mut report) {
-                Some(bytes) => {
-                    if bytes != case.wire {
-                        fail(&format!("{}: the wire round trip drifted", case.name));
-                    }
-                }
-                None => fail(&format!("{}: the wire round trip refused", case.name)),
-            }
-            match (case.codec.json_read)(&case.text, &mut report) {
-                Some(bytes) => {
-                    if bytes != case.wire {
-                        fail(&format!("{}: the text read no longer packs to its wire golden", case.name));
-                    }
-                }
-                None => fail(&format!("{}: the text read refused", case.name)),
-            }
-            match (case.codec.json_write)(&case.wire, &mut report) {
-                Some(text) => {
-                    if text != case.text {
-                        fail(&format!("{}: the text write drifted", case.name));
-                    }
-                }
-                None => fail(&format!("{}: the text write refused", case.name)),
-            }
-        }
-        // and the same gate every iteration, so a path that allocates on the
-        // hundredth pass and not the first is caught too
-        for case in cases.iter() {
-            let made = (case.codec.audit)(&case.wire, &case.text, &mut audit_buffer);
-            if made != 0 {
-                fail(&format!(
-                    "{}: the generated code made {made} allocation(s) on the read and write paths",
-                    case.name
-                ));
-            }
-        }
-        iterations += 1;
-        let now = Instant::now();
-        if now >= next_report {
-            let live = LIVE_BYTES.load(Ordering::Relaxed);
-            let drift = live - baseline;
-            println!(
-                "soak {:>5}s  {:>8} iterations  live {:>10} bytes  drift {:>+8}  allocations {}",
-                (now - started).as_secs(),
-                iterations,
-                live,
-                drift,
-                TOTAL_ALLOCATIONS.load(Ordering::Relaxed)
-            );
-            next_report = now + Duration::from_secs(60.min(seconds));
-        }
-    }
-
-    let live = LIVE_BYTES.load(Ordering::Relaxed);
-    let drift = live - baseline;
-    let allocations = TOTAL_ALLOCATIONS.load(Ordering::Relaxed) - allocations_baseline;
-    println!(
-        "soak: {iterations} iterations of {} instances over {seconds}s, live {live} bytes, drift {drift:+}",
-        cases.len()
-    );
-    println!(
-        "soak: the generated read and write paths made 0 allocations across all of it; \
-         this driver's own buffers made {allocations}"
-    );
-    if drift != 0 {
-        fail(&format!(
-            "the soak's live allocated bytes moved by {drift:+} — a read path that allocates does not stay flat"
-        ));
-    }
-}
-
-// THE ALLOCATION AUDIT as its own mode, so the number in the PR body is one
-// command rather than a paragraph. Every instance of the corpus, every
-// generated read and write path, counted at the global allocator with the
-// driver's own buffers hoisted out of the measured region.
-fn surface_alloc_audit(manifest: &Manifest) {
-    let rows = codecs();
-    let mut buffer: Vec<u8> = Vec::new();
-    let mut worst = 0u64;
-    println!("{:<24} {:>12}", "instance", "allocations");
-    for f in manifest.of_kind("instance") {
-        // the pointered class is ABSENT here (§11), and the deep chain has no
-        // text (§16.7): neither is a path this backend has to audit
-        if no_text(f) {
-            continue;
-        }
-        let Some(codec) = codec_for(&rows, &f[2], &f[3]) else {
-            continue;
-        };
-        let wire = slurp(&f[4]);
-        let text = slurp(&format!("testdata/conformance/tables/json/{}.json", f[1]));
-        // one warm call, then the measured one
-        (codec.audit)(&wire, &text, &mut buffer);
-        let made = (codec.audit)(&wire, &text, &mut buffer);
-        println!("{:<24} {made:>12}", f[1]);
-        worst = worst.max(made);
-    }
-    if worst != 0 {
-        fail("the generated code allocates on a read or write path");
-    }
-    println!(
-        "alloc audit: load, measure, save, from_json, to_json_measure and to_json — 0 allocations, every instance"
-    );
-}
-
 fn main() {
     let args: Vec<String> = std::env::args().collect();
     if args.len() < 3 {
@@ -1335,31 +783,17 @@ fn main() {
         );
         return;
     }
-    if surface == "alloc-audit" {
-        surface_alloc_audit(&manifest);
-        return;
-    }
     if args.len() < 4 {
         eprintln!("usage: {} <manifest> <surface> <outdir>", args[0]);
         exit(2);
     }
     let out = args[3].as_str();
     match surface {
-        "wire" => surface_wire(&manifest, out),
-        "report" => surface_report(&manifest, out),
-        "json-read" => surface_json_read(&manifest, out),
-        "json-write" => surface_json_write(&manifest, out),
-        "json-hostile" => surface_json_hostile(&manifest, out),
         "cook" => surface_cook(&manifest, out),
         "cook-foreign" => surface_cook_foreign(&manifest, out),
         "block-foreign" => surface_block_foreign(&manifest, out),
         "block-dump" => surface_block_dump(&manifest, out),
         "cook-forgery" => surface_cook_forgery(&manifest, out),
-        // NOT a conformance surface: the harness never asks for it, and `list`
-        // does not name it. It is here because this binary already carries
-        // every codec of the corpus, exactly as the C++ driver carries its
-        // block-forgery pinning mode.
-        "soak" => surface_soak(&manifest, out.parse::<u64>().unwrap_or(3600)),
         "block" => surface_block(&manifest, out),
         "forgery" => surface_forgery(&manifest, out),
         _ => exit(2),
